refactor(ai): move the 22 prompt constants to src/analysis/ai/prompts/
pipeline.ts 6,136 -> 4,883 lines. 1,253 lines of prompt text now live in seven
files under src/analysis/ai/prompts/, none larger than 355 lines.

THE MOVE IS BYTE-EXACT, AND THAT IS MECHANICALLY PROVEN. tests/eval/changeGate.ts
hashes SYSTEM_PROMPT, CSV_SYSTEM_PROMPT, LOG_SYSTEM_PROMPT and SYNTHESIS_PROMPT
and demands a real-model no-regression attestation whenever the hash moves (#378).
After this commit it reports:

  AI evaluation change gate: not-required -- default evaluation prompts and
  models are unchanged

That is the whole argument for how this was done. The blocks were sliced out
programmatically and each of the 22 verified byte-identical to the original
before anything was deleted -- because a single reflowed line is
indistinguishable, to that hash, from an edited prompt, and would have demanded
an attestation for a change that is not one.

THE SPLIT IS BY WHAT THE PROMPT IS FOR, since that is how someone looking for
one will search: extraction (vision/CSV/log/observe/importer), synthesis
(synthesis + the two narrative renderings), hunt, investigation (ask, hypothesis
review/generation, next step, explain), findings (tagger rules, FP similarity,
remediation) and reporting (starred, view, session). index.ts owns the one thing
that must know about all of them at once -- resolvePrompt, which resolves a
built-in against its DFIR_AI_<NAME>_PROMPT[_FILE] override. Splitting that per
theme would scatter the override contract across six files and make "what can an
operator override?" unanswerable from any single place.

FOUR CONSUMERS READ THE PROMPTS BY PATH, and all four are updated:
changeGate.ts, checkChange.ts, identity.ts, eject-prompts.ts. changeGate.ts
gains collectPromptSource(), which reads the new layout and FALLS BACK to
pipeline.ts. The fallback is not defensive clutter: baseSourceHash() reads the
merge base, and every revision before this commit still has the prompts in
pipeline.ts, so without it the gate would report a spurious prompt change on
every PR until this lands on master.

A LATENT .gitignore BUG SWALLOWED THE NEW DIRECTORY. `prompts/` was unanchored,
so it matched ANY directory of that name at any depth -- including
src/analysis/ai/prompts/. The first commit of this change was broken: pipeline.ts
imported seven modules that git was silently refusing to track, so a fresh clone
would not have built. Anchored to /prompts/, which still ignores the
`npm run prompts:eject` output it was written for. The file already carried the
same fix for /velociraptor/, with a comment explaining why -- this is the second
instance of that trap.

pipeline.ts re-exports the whole prompt surface, so the 23 modules importing
these names from it are untouched.

Verified: build, typecheck, lint, format:check, check:size (pipeline.ts
re-recorded), check:imports, check:boundaries, eval:change-gate, and the full
suite -- 531 files, 6,746 tests, all passing.

Refs #384
