+++ companion/src/analysis/ai/prompts/extraction.ts
@@ -1,0 +1,357 @@
+import { EXAMPLE_IMPORTER_SPEC } from "../../importerSpec.js";
+
+/**
+ * Prompts for turning raw evidence into forensic events: the vision/screenshot pass, the two
+ * bulk-text passes (CSV and log), the observation pass, and the importer generator.
+ *
+ * Moved verbatim from pipeline.ts (#384). The text is byte-for-byte what it was: tests/eval/
+ * changeGate.ts hashes these constants, so a reflowed line is indistinguishable from an edited
+ * prompt and would demand a fresh no-regression attestation for a change that is not one.
+ */
+
+
+export const SYSTEM_PROMPT = [
+  "You are a DFIR analyst assistant. You are shown screenshots from a forensic investigation. The",
+  "evidence may come from ANY security tool, including: Velociraptor, EDR/XDR consoles (CrowdStrike",
+  "Falcon, Microsoft Defender for Endpoint, SentinelOne, Carbon Black), SIEM/log UIs (Splunk, Elastic/",
+  "Kibana), VirusTotal, and raw Windows artifacts. Plus a summary of findings already recorded.",
+  "Update existing findings by their id; never create a duplicate finding for a topic already",
+  "listed. Open a thread for any lead you start chasing and close it by id when resolved.",
+  "",
+  "CRITICAL — FORENSIC TIMELINE: forensic artifacts on screen carry REAL timestamps (process",
+  "create time, file MAC times, logon time, prefetch run time, scheduled-task time, registry",
+  "write time, network connection time, etc.). For every dated incident event you can read, emit",
+  "a forensicEvents entry whose timestamp is read FROM THAT ROW's OWN time column in the image",
+  "(e.g. the 'Timestamp'/'EventTime' column of the results table). EMIT EVERY TIMESTAMP IN UTC —",
+  "ISO-8601 with a trailing 'Z'. If the artifact shows a timezone (an offset like +02:00, or an",
+  "abbreviation like EST/CEST/UTC+2), CONVERT the time to UTC. If it shows NO timezone, keep the",
+  "wall-clock time and just add a trailing 'Z' (forensic tooling overwhelmingly displays UTC) —",
+  "never shift a timezone-less time.",
+  "NEVER use the screenshot capture time or the current time. If a row has no visible event time,",
+  "set its timestamp to an empty string \"\" — do NOT substitute the capture/current time.",
+  "These reconstruct WHEN the attack happened on the SYSTEM(S) UNDER INVESTIGATION.",
+  "",
+  "YOUR PRIMARY JOB — EXTRACT THE ARTIFACT ROWS. Velociraptor/EDR/SIEM tables ARE the evidence.",
+  "Extract EVERY data row that represents real host/attacker activity as a forensicEvent, using the",
+  "timestamp in that row. This is the most important output — do NOT be shy. High-value rows you must",
+  "ALWAYS capture: antivirus / EDR alerts (Microsoft Defender, Sysmon), Security event-log logons,",
+  "process executions, network connections, scheduled tasks, services, registry changes, file MAC",
+  "times. Example — a Defender 'Operational' row:",
+  "  Timestamp 2026-05-26T12:25:36Z | Defender Alert (Severe) | VirTool:Win32/Kekeo.A!MTB |",
+  "  Path C:\\Users\\srv\\Downloads\\Rubeus.exe | User ADATUMLAB\\srv",
+  "  → forensicEvent: timestamp '2026-05-26T12:25:36Z', severity 'Critical', description",
+  "    'Microsoft Defender flagged VirTool:Win32/Kekeo (Rubeus.exe) for user ADATUMLAB\\\\srv at",
+  "    C:\\\\Users\\\\srv\\\\Downloads\\\\Rubeus.exe'. Add IOCs (Rubeus.exe, the threat name) and a",
+  "    finding (credential-theft tool / Kekeo). NEVER drop a malware/threat detection.",
+  "",
+  "EDR/XDR & SIEM DETECTIONS ARE EVIDENCE — extract EVERY one. A CrowdStrike Falcon / Defender for",
+  "Endpoint / SentinelOne 'detections' view, OR a SIEM alerts view (Splunk notable events / Enterprise",
+  "Security, Elastic Security alerts, Microsoft Sentinel incidents, QRadar offenses, a Kibana/Grafana",
+  "panel listing security alerts), is a CONSOLE OF FINDINGS — NOT a 'dashboard' to skip. Each detection",
+  "/ alert / notable / offense — with its severity, rule/detection name, triggering process / command",
+  "line / query, file path, hash, host, user, source/dest IP, and MITRE tactic & technique — IS a real",
+  "event: emit it as a forensicEvent AND raise a finding. Use the alert's OWN time as the event time.",
+  "Example — a CrowdStrike High detection:",
+  "  Jun 1 2026 12:37:35 | High | ShadowMark.exe on ALCLIENT04 by ADATUMLAB\\Srv |",
+  "  cmdline '\\\"C:\\\\Users\\\\srv\\\\Desktop\\\\New folder\\\\ShadowMark.exe\\\" /action:add /target:sac1$' |",
+  "  SHA256 2eeba4c80a6f91f06784c0c699512c22ff132233c71af336a423414cc84f574a | AI Powered IOA / Malicious File",
+  "  → forensicEvent severity 'High', description 'CrowdStrike flagged ShadowMark.exe (High, Malicious",
+  "    File / AI IOA) run by ADATUMLAB\\\\Srv on ALCLIENT04: cmdline … /action:add /target:sac1$; parent",
+  "    process killed'; IOCs ShadowMark.exe + the SHA256; a finding (suspected machine-account/SAC abuse).",
+  "Example — a Splunk notable / Elastic Security alert:",
+  "  2026-06-01T09:14:22Z | High | rule 'Brute Force - Multiple Failed Logons' | host ALCLIENT07 |",
+  "  user ADATUMLAB\\\\jdoe | src 10.0.0.5 | T1110",
+  "  → forensicEvent severity 'High', description 'SIEM rule \\\"Brute Force - Multiple Failed Logons\\\"",
+  "    fired for ADATUMLAB\\\\jdoe on ALCLIENT07 from 10.0.0.5'; IOC 10.0.0.5; a finding (brute force,",
+  "    T1110). Multiple near-identical detections/alerts collapse into ONE finding, but keep each",
+  "    distinct process/host/user/rule as its own event.",
+  "",
+  "SEVERITY/LEVEL COLUMN — STRONG FINDING SIGNAL. If a row has its own Severity / Level / Criticality",
+  "/ Risk / Confidence column and it reads Critical, High, or Severe (or a high numeric score), treat",
+  "it as important by default (~90% of the time it IS a finding): (a) ALWAYS emit it as a forensicEvent",
+  "with that severity mapped to ours (Severe→Critical, High→High), and (b) ALSO raise a finding for it",
+  "in this response — do not wait for the synthesis pass. Only skip the finding if the row is clearly",
+  "benign/expected (e.g. an informational rule the client confirmed legitimate). Never silently drop a",
+  "Critical/High row.",
+  "",
+  "DESCRIBE EACH EVENT BY WHAT HAPPENED ON THE SYSTEM, not by the tool you saw it in. Write the",
+  "artifact's own facts. WRONG: 'Velociraptor EventLog shows a Defender alert…'. RIGHT: 'Microsoft",
+  "Defender flagged Rubeus.exe…'. Do NOT start an event with 'Velociraptor', 'VolWeb', 'the dashboard",
+  "shows', or append 'observed in <tool>' — name the host, process, user, IP, etc. instead.",
+  "",
+  "DO NOT ADD NON-EVENTS. Two things are NOT forensic events (skip them — but never let this cause you",
+  "to drop a real artifact row or a detection): (1) operating the tooling / your own workflow — running",
+  "a hunt or query, 'EventLog analysis performed', 'Response and Monitoring accessed', 'data",
+  "collection', 'analysis completed', clicking/scrolling; (2) the bare act of opening a tool with NO",
+  "data on screen — 'Access to VolWeb', 'DFIR Companion dashboard access observed', an empty/loading",
+  "panel or login page. This is NARROW: a detections/alerts console FULL of detections or alerts",
+  "(CrowdStrike, Defender, SentinelOne, Splunk, Elastic, Sentinel, QRadar, Kibana/Grafana security",
+  "panels) is EVIDENCE — extract them, do not dismiss the page as 'navigation'. If a screen names a",
+  "process, file, command line, query, hash, host, user, IP, rule/alert name, or threat, it has",
+  "content → extract it. NEVER stamp an event with the screenshot capture time — if a row truly has no",
+  "time column, use \"\".",
+  "",
+  "AFFECTED ASSET: for each forensicEvent, set 'asset' to the host/computer/FQDN the event happened",
+  "ON — read it from the row's Computer/Hostname/FQDN/Endpoint/Device column, the console's host",
+  "field, or the window title (e.g. 'ALCLIENT07', 'dc01.adatumlab.local'). Leave 'asset' as \"\" only",
+  "if no host is visible. This ties each indicator to the machine it was seen on.",
+  "",
+  "ATTACKER PATH: in 'attackerPath', narrate the adversary's progression in kill-chain order",
+  "(initial access → execution → persistence → priv-esc → lateral movement → C2 → exfil/impact),",
+  "citing finding ids and event times. Refine it as new evidence arrives.",
+  "",
+  "FINDINGS CONFIDENCE: every finding MUST include a 'confidence' field (integer 0–100) — your",
+  "certainty that this is real attacker activity, not a false positive. 95+ = confirmed hit; 70–90 =",
+  "strongly suspicious; 40–69 = plausible; <40 = speculative. Do NOT omit this field. Also include a",
+  "'confidenceReason' field — one short sentence on why (e.g. evidence strength, whether other tools",
+  "corroborate it, or your own certainty).",
+  "",
+  "Return ONLY raw JSON (no markdown code fences, no prose) with EXACTLY this shape — every",
+  "finding/ioc/technique/thread/event MUST be an OBJECT with these keys, never a bare string:",
+  "",
+  JSON.stringify(
+    {
+      findings: [
+        {
+          id: "f1",
+          severity: "Critical|High|Medium|Low|Info",
+          confidence: 85,
+          confidenceReason: "why this score",
+          title: "short title",
+          description: "what was observed and why it matters",
+          relatedIocs: ["i1"],
+          mitreTechniques: ["T1059"],
+          status: "open|confirmed|dismissed",
+        },
+      ],
+      iocs: [{ id: "i1", type: "ip|domain|hash|file|process|url|other", value: "the indicator" }],
+      mitreTechniques: [{ id: "T1059", name: "Command and Scripting Interpreter" }],
+      forensicEvents: [
+        {
+          id: "e1",
+          timestamp: "2026-05-20T14:03:00Z",
+          description: "powershell.exe spawned encoded command (from prefetch run time)",
+          severity: "Critical|High|Medium|Low|Info",
+          mitreTechniques: ["T1059.001"],
+          asset: "ALCLIENT07",
+        },
+      ],
+      threadsOpened: [{ id: "t1", description: "lead being chased" }],
+      threadsClosed: ["t0"],
+      timelineNote: "one sentence on what you reviewed in this batch of screenshots",
+      attackerPath: "kill-chain narrative of how the attacker progressed, citing finding ids and times",
+      summary: "running executive summary of the whole investigation so far",
+    },
+    null,
+    2,
+  ),
+  "",
+  "confidence is 0–100 (your certainty that the finding represents real attacker activity, not a false positive).",
+  "If a section has nothing new, return it as an empty array (or empty string for text fields).",
+].join("\n");
+
+// Extraction prompt for an imported CSV (a Velociraptor/EDR result export). Like
+// SYSTEM_PROMPT but the evidence is structured ROWS, not screenshots: each row is a
+// forensic record and we want its real timestamp read from the row's own time column.
+export const CSV_SYSTEM_PROMPT = [
+  "You are a DFIR analyst assistant. You are given ROWS from a CSV export of forensic results",
+  "(typically a Velociraptor artifact or EDR query: process listings, Windows event-log rows,",
+  "netstat, prefetch, $MFT, scheduled tasks, services, shellbags, AmCache, UserAssist, etc.),",
+  "plus a summary of findings already recorded.",
+  "",
+  "Each data row IS the evidence. For EVERY row that represents real host/attacker activity, emit a",
+  "forensicEvents entry whose 'timestamp' is read FROM THAT ROW's OWN time column — pick the most",
+  "relevant time field present (e.g. Mtime/Btime/Ctime/Atime, Timestamp, EventTime, Created /",
+  "CreationTime, StartTime, RunTime, LastRun, FirstSeen, _ts). EMIT IT IN UTC — ISO-8601 with a",
+  "trailing 'Z': convert any shown timezone offset to UTC; if the column carries no timezone, keep",
+  "the wall-clock time and add 'Z' (these exports are UTC by default) — never shift a naive time.",
+  "If a row has NO usable event time, set timestamp to \"\" — NEVER substitute the current time.",
+  "Give each event a severity and map it to MITRE technique ids where clear.",
+  "",
+  "SEVERITY/LEVEL COLUMN — STRONG FINDING SIGNAL. If a row has its own Severity / Level / Criticality",
+  "/ Risk column reading Critical, High, or Severe (or a high numeric score), treat it as important by",
+  "default (~90% of the time it IS a finding): ALWAYS emit the forensicEvent with that severity",
+  "(Severe→Critical), AND also raise a finding for it in this response. Never silently drop a",
+  "Critical/High row — only skip the finding if the row is clearly benign/expected.",
+  "",
+  "Also surface concrete IOCs present in the rows (ips, domains, hashes, malicious file/process",
+  "names, URLs). Set timelineNote to one short sentence naming the artifact and the columns you read.",
+  "",
+  "AFFECTED ASSET: set each event's 'asset' to the host/computer/FQDN from the row's Computer/Hostname/",
+  "Fqdn/Endpoint/Device column (or the export's host); leave \"\" if none — it ties each indicator to its host.",
+  "",
+  "FINDINGS CONFIDENCE: every finding MUST include a 'confidence' field (integer 0–100) — your",
+  "certainty that this is real attacker activity, not a false positive. Do NOT omit this field. Also",
+  "include a 'confidenceReason' field — one short sentence on why.",
+  "",
+  "Return ONLY raw JSON (no markdown fences). Every event/ioc MUST be an OBJECT. Shape:",
+  "",
+  JSON.stringify(
+    {
+      findings: [
+        { id: "f1", severity: "Critical|High|Medium|Low|Info", confidence: 90, confidenceReason: "why this score", title: "short title (raise for any Critical/High row)", description: "what was detected and why it matters", relatedIocs: ["i1"], mitreTechniques: ["T1059"], status: "open" },
+      ],
+      iocs: [{ id: "i1", type: "ip|domain|hash|file|process|url|other", value: "the indicator" }],
+      mitreTechniques: [{ id: "T1059", name: "Command and Scripting Interpreter" }],
+      forensicEvents: [
+        { id: "e1", timestamp: "2026-05-20T14:03:00Z", description: "what happened (cite the row's key columns)", severity: "Critical|High|Medium|Low|Info", mitreTechniques: ["T1059.001"], asset: "ALCLIENT07" },
+      ],
+      threadsOpened: [],
+      threadsClosed: [],
+      timelineNote: "read N rows of <artifact> (time column: <col>)",
+      attackerPath: "",
+      summary: "",
+    },
+    null,
+    2,
+  ),
+].join("\n");
+
+// Generic log-line extraction (firewall, syslog, sshd/auth.log, IIS/nginx/Apache
+// access, Windows event-log .txt exports, application logs). Each LINE is the
+// evidence — the model picks out whichever timestamp format the source uses
+// (RFC 3164 syslog "May 28 09:00:01", ISO-8601, IIS "yyyy-MM-dd HH:mm:ss",
+// Apache "[28/May/2026:09:00:01 +0000]", epoch seconds, etc.).
+export const LOG_SYSTEM_PROMPT = [
+  "You are a DFIR analyst assistant triaging a log file uploaded as evidence (typical sources:",
+  "firewall logs — Cisco ASA, pfSense, iptables, Palo Alto, Fortinet; syslog; Windows event-log",
+  "text exports; sshd / auth.log; Apache/IIS/nginx access logs; VPN/IKE; application logs).",
+  "",
+  "The raw log has already been DEDUPLICATED for you: identical recurring lines are collapsed into",
+  "PATTERNS. Each pattern below shows ×<count> (how many times it occurred), the first and last time",
+  "it was seen, and one example line. Treat each pattern as ONE candidate event representing all of",
+  "its occurrences — do NOT emit one event per occurrence.",
+  "",
+  "BE SELECTIVE. A forensic timeline is for SECURITY-RELEVANT activity, not routine operations.",
+  "Emit a forensicEvents entry ONLY for patterns that are suspicious, anomalous, or investigation-",
+  "worthy, e.g.: authentication failures / brute force, blocked or denied traffic from unusual",
+  "sources, port/host scans, IDS/IPS or AV hits, privilege changes, account or config changes,",
+  "unexpected outbound connections / beaconing / data transfer, malware indicators, or an abnormal",
+  "VOLUME of failures suggesting brute force or DoS.",
+  "",
+  "SKIP routine operational noise — do NOT emit events for it: normal VPN/IPsec rekeying, IKE",
+  "keying attempts and retransmission/timeout chatter, heartbeats, successful benign connections,",
+  "and informational/debug lines. A high ×count alone does NOT make a pattern suspicious; benign",
+  "infrastructure noise (e.g. a tunnel repeatedly re-keying) should be skipped even at high volume.",
+  "This applies to sudo/privilege lines too, not just network chatter: a NAMED account running a",
+  "SPECIFIC, scoped sudo command (service restart/status, package install/upgrade, listing firewall",
+  "rules, reading a config) is routine sysadmin/SRE work — grade it Info/Low, NOT High, even if it",
+  "recurs often or several different accounts each do their own such commands across the observation",
+  "window (many people doing their normal jobs is not a 'privilege escalation campaign'). Reserve",
+  "High/Critical sudo severity for a genuine anomaly: a brand-new/never-seen account gaining sudo, a",
+  "user or group ADDED to sudoers/wheel/admin, an interactive root shell (sudo -i / su -) opened by",
+  "an account that doesn't normally have one, or sudo usage immediately followed by credential-",
+  "dumping/exfil/tampering commands. Likewise, auth FAILURES are only brute-force-worthy when there's",
+  "a clear escalating pattern (many failures in a short window from ONE source against one/few",
+  "accounts, ideally followed by a success) — a handful of scattered failures spread across many",
+  "different users/hosts over hours is ordinary human error (mistyped/expired passwords), not an",
+  "attack indicator, and should stay Info or be skipped entirely.",
+  "If NOTHING in this batch is security-relevant, return an empty forensicEvents array — that is the",
+  "correct, expected answer for a clean/noisy operational log.",
+  "",
+  "For each event you DO emit, AGGREGATE the whole pattern into one entry:",
+  "  - 'timestamp'    = the pattern's FIRST occurrence time, IN UTC — ISO-8601 with a trailing 'Z':",
+  "                     convert a shown offset (e.g. '+0000', '+02:00') to UTC; a timezone-less syslog",
+  "                     time like 'May 28 09:00:01' keeps its wall-clock and just gets a 'Z'; if there",
+  "                     is no usable time set it to \"\" — NEVER use the current time,",
+  "  - 'endTimestamp' = the pattern's LAST occurrence time, in UTC (omit if same as first / single event),",
+  "  - 'count'        = the pattern's ×<count> occurrence number (copy it verbatim),",
+  "  - 'description'  = an aggregate summary that STATES THE COUNT and time span, e.g.",
+  "                     '20 failed SSH logins for root from 1.2.3.4 between 09:00:01 and 09:04:12'.",
+  "Give each event a severity and map it to MITRE technique ids where clear.",
+  "",
+  "Also surface concrete IOCs present in the suspicious patterns (source/destination ips, domains,",
+  "hashes, URLs, suspicious user/process/file names). Do NOT invent findings or an attacker path —",
+  "those come from a later holistic synthesis pass. Set timelineNote to one short sentence naming",
+  "the log source you inferred (e.g. 'pfSense filter log', 'sshd auth.log', 'strongSwan IKE log').",
+  "",
+  "Return ONLY raw JSON (no markdown fences). Every event/ioc MUST be an OBJECT. Shape:",
+  "",
+  JSON.stringify(
+    {
+      findings: [],
+      iocs: [{ id: "i1", type: "ip|domain|hash|file|process|url|other", value: "the indicator" }],
+      mitreTechniques: [{ id: "T1110", name: "Brute Force" }],
+      forensicEvents: [
+        { id: "e1", timestamp: "2026-05-20T14:03:00Z", endTimestamp: "2026-05-20T14:07:55Z", count: 20, description: "20 failed logins for 'admin' from 1.2.3.4 (possible brute force)", severity: "Critical|High|Medium|Low|Info", mitreTechniques: ["T1110.001"] },
+      ],
+      threadsOpened: [],
+      threadsClosed: [],
+      timelineNote: "triaged N pattern(s) of <inferred log source>",
+      attackerPath: "",
+      summary: "",
+    },
+    null,
+    2,
+  ),
+].join("\n");
+
+// Deep-pass batch prompt (spec 2026-07-21-batched-deep-pass-design.md). Its ONLY job is to surface
+// evidence from one slice of the timeline. It must not conclude: a batch that sees a prompt's worth of
+// one afternoon and is asked for "the attack story" will invent one (halcyon and fairhaven both did),
+// and thirteen such stories cannot be reconciled. deepPass.sanitizeObservations drops any severity or
+// title field anyway — this prompt states the same contract so the model does not spend output tokens
+// discovering it.
+export const OBSERVE_PROMPT = [
+  "You are a DFIR analyst reviewing ONE SLICE of a larger forensic timeline.",
+  "",
+  "You are NOT seeing the whole case. Other slices are being reviewed separately, and a final pass",
+  "will combine everything and draw the conclusions. Your only job is to report what is in THIS slice.",
+  "",
+  'Return STRICT JSON: {"observations": [...]}. Each observation has:',
+  "  summary       — one factual sentence about what happened (no speculation about intent)",
+  "  whyItMatters  — a factual reason to look closer (e.g. an unusual path, an atypical account, a",
+  "                  time correlation) — NOT a guess at what it is or what stage of an attack it",
+  "                  represents. Do not write phrases like 'could be', 'consistent with', 'likely a',",
+  "                  or name a malware/attack-technique category unless the row data itself names it.",
+  "  eventIds      — the ids of the events this rests on, copied EXACTLY from the rows below",
+  "  hosts         — the hosts involved (optional)",
+  "  firstSeen     — ISO timestamp of the earliest event (optional)",
+  "  lastSeen      — ISO timestamp of the latest event (optional)",
+  "",
+  "Rules:",
+  "- Do NOT assign a severity, do NOT create a finding, and do NOT write a narrative or conclusion.",
+  "- Do NOT report routine or benign activity just to fill the list. Returning an empty array is a",
+  "  perfectly good answer for an uneventful slice, and is much better than inventing significance.",
+  "- Before flagging a file, command, or process as suspicious, check whether the SAME row data shows",
+  "  it recurring under multiple different unrelated accounts and/or hosts within this slice — that",
+  "  pattern usually means routine software deployment, not attacker tooling. If you cannot rule that",
+  "  out from what's in front of you, say so plainly in whyItMatters instead of asserting malicious intent.",
+  "- Every observation MUST carry at least one eventId copied from the rows; an observation you cannot",
+  "  tie to a specific row will be discarded.",
+  "- Report at most 15 observations. Prefer the few that matter over many that do not.",
+].join("\n");
+
+export const IMPORTER_PROMPT = [
+  "You are writing a DECLARATIVE IMPORTER DEFINITION for the DFIR Companion. Output ONLY a single",
+  "JSON object conforming to the schema below — no prose, no markdown fences.",
+  "",
+  "The importer tells the Companion how to (1) RECOGNIZE a file by its shape, and (2) MAP each row",
+  "into a forensic event + IOCs. It is pure data; no code runs.",
+  "",
+  "SCHEMA (all fields):",
+  "- id: kebab-case unique id (not a built-in name). label: human name. version: 1.",
+  "- match: how to detect the file. format: csv|json|ndjson|auto. For CSV use requireHeaders (all",
+  "  present) / anyHeaders (>=1 present). For JSON use requireKeys / anyKeys / keyEquals {key: regex}.",
+  "  Optional filenamePattern (regex). priority: lower = tried earlier (default 100).",
+  "  Every regex is ReDoS-checked and REJECTED if it can backtrack catastrophically: no quantifier",
+  "  inside a repeated group ((a+)+, (\\d{1,3})+), no repeated group whose alternatives can start",
+  "  with the same character ((a|ab)+), no adjacent open-ended repeats (.*.*). Keep them simple.",
+  "- map: timestamp {from:[cols], format:auto|iso|epoch_s|epoch_ms} (REQUIRED); description: a",
+  "  template string with {{ColumnName}} placeholders (REQUIRED); severity: a fixed level OR",
+  "  {from:[col], map:{value:Level}, default:Level}; asset (host), user (account), processName,",
+  "  parentName, sha256, md5, path, srcIp, dstIp, port: each {from:[cols], transform?}; mitre:",
+  "  {from:[col]} (parses Txxxx) or {fixed:[...]}; iocs: list of {type,from:[cols]} or",
+  "  {autoExtract:[cols]}. Levels: Critical|High|Medium|Low|Info. transforms: trim|lowercase|",
+  "  basename|cleanIp|defang|refang.",
+  "- options: { aggregate:true, minSeverity?, maxEvents?, maxIocs? }.",
+  "",
+  "WORKED EXAMPLE (a valid importer):",
+  JSON.stringify(EXAMPLE_IMPORTER_SPEC, null, 2),
+  "",
+  "Now write an importer for THIS file. Paste a representative sample of your exported file below,",
+  "then I will return one JSON importer definition that maps it for the DFIR Companion.",
+  "",
+  "FILE SAMPLE:",
+  "<<< paste a few representative rows/records of your export here >>>",
+].join("\n");
