+++ companion/src/analysis/ai/prompts/hunt.ts
@@ -1,0 +1,205 @@
+/**
+ * Prompts that generate hunts and translate analyst questions into the internal query language.
+ *
+ * Moved verbatim from pipeline.ts (#384). The text is byte-for-byte what it was: tests/eval/
+ * changeGate.ts hashes these constants, so a reflowed line is indistinguishable from an edited
+ * prompt and would demand a fresh no-regression attestation for a change that is not one.
+ */
+
+
+// Propose PROACTIVE Velociraptor VQL fleet-hunts from the synthesized findings (issue #57). The
+// model reads the findings / ATT&CK techniques / pivotable IOCs and emits hunts that run on EVERY
+// enrolled endpoint to find the same tradecraft elsewhere. The VQL must be deployable as-is via the
+// CLIENT-artifact hunt path (launchHunt), so the shape constraints here mirror what splitVqlStatements
+// expects: ONE self-contained VQL statement per hunt, no blank-line splits, not comment-only.
+export const HUNT_SUGGEST_PROMPT = [
+  "You are a senior DFIR threat hunter. Given ONE investigation's findings, ATT&CK techniques,",
+  "compromised assets, and indicators below, propose PROACTIVE Velociraptor VQL HUNTS that run",
+  "across the ENTIRE fleet of enrolled endpoints to find the SAME adversary tradecraft on hosts",
+  "that are not yet in scope — lateral spread, the same webshell/persistence/malware pattern,",
+  "the same C2, the same living-off-the-land technique.",
+  "",
+  "Rules:",
+  "- Hunt for the PATTERN across the fleet, do NOT merely restate a single-host finding. If a",
+  "  webshell was found in one web root, hunt every host's web roots; if a malicious service was",
+  "  installed, enumerate that service/registry value everywhere.",
+  "- When an ATTACK GRAPH is shown below, hunt for the RELATIONSHIP it reveals — the parent→child",
+  "  spawn chain, the wrote→executed file lineage, the binary/account that moved between hosts — not",
+  "  just the leaf indicator alone. A chain (e.g. winword.exe → powershell.exe → rundll32.exe) is far",
+  "  more specific than any single process name, and a binary seen on two hosts means hunt the rest.",
+  "- Pivot ONLY on the case's REAL indicators (the exact hashes, file paths, process names,",
+  "  service names, domains, IPs shown below). Do NOT invent IOCs the case does not contain.",
+  "- Each `vql` MUST be a SINGLE, self-contained, CLIENT-side Velociraptor VQL statement that runs",
+  "  on each endpoint — one `SELECT … FROM <plugin>(…) WHERE …`. Use real Velociraptor plugins,",
+  "  e.g. glob(), stat(), pslist(), Artifact.Windows.System.Services, Artifact.Windows.Sys.Users,",
+  "  read_file(), hash(), yara(), Artifact.Windows.Registry.* . Velociraptor glob() uses FORWARD",
+  "  slashes. Do NOT put a blank line inside one query and do NOT make a query only a comment.",
+  "- Use each plugin's REAL argument names. parse_evtx() takes `filename=` (a path or glob), NOT",
+  "  `files=` — e.g. parse_evtx(filename='C:/Windows/System32/winevt/Logs/Security.evtx'). For Windows",
+  "  event-log hunts PREFER Velociraptor's own artifacts (Windows.EventLogs.* / Windows.Detection.*)",
+  "  over hand-writing parse_evtx, and access fields as `System.EventID.Value` / `EventData.<Name>`.",
+  "- read_file() takes `filenames=` (a LIST), NOT `filename=` (scalar). To read one file:",
+  "  `SELECT Data FROM read_file(filenames=['C:/path/file.txt'])`. To read files found by glob,",
+  "  call it inline in the SELECT (no JOIN needed):",
+  "  `SELECT FullPath, read_file(filenames=[FullPath])[0].Data AS Content FROM glob(globs='...')`.",
+  "- VQL has NO SQL `JOIN`. Use inline function calls (e.g. read_file above) or",
+  "  `foreach(row={SELECT … FROM a()}, query={SELECT … FROM b()})` to correlate two plugins.",
+  "- hash() takes ONLY `path=` and returns an object with `.MD5` / `.SHA1` / `.SHA256` — there is NO",
+  "  `hashselect`/`algorithm`/`type` argument (inventing one fails to COMPILE, so the hunt never starts).",
+  "  Get a file's SHA256 as `hash(path=FullPath).SHA256`, computed ONCE as a column, not repeatedly in WHERE:",
+  "  `SELECT FullPath, hash(path=FullPath).SHA256 AS SHA256 FROM glob(globs='…/*.exe') WHERE SHA256 IN (…)`.",
+  "- NEVER hash with a full-disk glob like 'C:/**/*.exe' — it walks every file on the volume and times out.",
+  "  Scope globs to the directories the tradecraft uses (web roots, %TEMP%, Downloads, a service path).",
+  "- PREFER raw VQL plugins — pslist(), netstat(), glob(), stat(), read_file(), hash(), yara() — over",
+  "  `Artifact.<Name>()` references: a hallucinated Artifact.<Name> fails to COMPILE and the hunt never",
+  "  starts. Network connections → netstat(); processes → pslist(); files → glob(). Only reference an",
+  "  Artifact.<Name> you are sure exists on the server.",
+  "- For an ABSOLUTE time use timestamp(string='2025-03-14T22:00:00Z'); epoch= takes unix SECONDS (a number).",
+  "- Velociraptor VQL has NO duration-suffix literals. Do NOT write `30d`, `7h`, `2w` etc.",
+  "  Use seconds arithmetic instead: `now() - 30 * 86400` (30 days), `now() - 7 * 86400` (7 days),",
+  "  `now() - 3600` (1 hour). Wrap in `timestamp(epoch=...)` when comparing against a timestamp",
+  "  column, e.g. `WHERE Mtime > timestamp(epoch=now() - 30 * 86400)`.",
+  "- Prefer a few HIGH-SIGNAL hunts over many near-duplicates. Skip a finding if there is nothing",
+  "  fleet-wide to hunt for it.",
+  "- For each hunt set: a short `title`; a `rationale` (which finding triggered it, what the query",
+  "  looks for, and how to triage a hit); `severity` (Critical|High|Medium|Low|Info) of the",
+  "  underlying threat; `mitreTechniques` (the finding's technique ids); and `relatedFindingIds`",
+  "  (the finding ids it derives from, using the [ids] shown).",
+  "",
+  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
+  JSON.stringify({
+    suggestions: [{
+      title: "Hunt for ASPX webshells across IIS web roots",
+      rationale: "Finding f3 shows an ASPX webshell on WEB01. Sweep every endpoint's web roots for .aspx files written recently; triage any hit by author/last-write time and contents.",
+      vql: "SELECT FullPath, Mtime, Size FROM glob(globs='C:/inetpub/wwwroot/**/*.aspx') WHERE Mtime > timestamp(epoch=now() - 2592000)",
+      severity: "High",
+      mitreTechniques: ["T1505.003"],
+      relatedFindingIds: ["f3"],
+    }],
+  }, null, 2),
+].join("\n");
+
+// Issue #70 — AI-suggested Velociraptor hunts for the case's PLAYBOOK tasks. For each ENDPOINT-related
+// task the model writes one CLIENT-side VQL hunt + echoes the single host it's scoped to (chosen ONLY
+// from the provided known-endpoints list). The server decides hunt-vs-collection deterministically
+// from the observed endpoints, so the model just needs to flag endpoint-relatedness and name the host.
+export const PLAYBOOK_HUNT_PROMPT = [
+  "You are a senior DFIR threat hunter. Below is ONE investigation's Response PLAYBOOK (the analyst's",
+  "actionable checklist), the case's known ENDPOINTS, findings, ATT&CK techniques, and forensic timeline.",
+  "For each playbook task that is ABOUT ENDPOINTS (collecting from, examining, containing, or hunting on",
+  "hosts — e.g. pull event logs, enumerate persistence, find a process/file, scope lateral movement),",
+  "propose ONE Velociraptor VQL hunt that gathers the evidence the task needs.",
+  "",
+  "Rules:",
+  "- Only emit a suggestion for a task that is genuinely endpoint-related. Set `endpointRelated` true for",
+  "  those and SKIP the rest (notify legal, rotate cloud creds, draft a report, block a domain at the",
+  "  firewall, etc. are NOT endpoint tasks). It is fine to return fewer suggestions than tasks.",
+  "- `taskId` MUST be the exact id (the [bracketed] value) of the task the suggestion is for.",
+  "- `targetHost`: if the task is about exactly ONE specific endpoint, set this to that host — but ONLY a",
+  "  hostname that appears in the KNOWN ENDPOINTS list below (the server runs it as a COLLECTION on just",
+  "  that one client). If the task spans multiple hosts, or you are unsure which host, set `targetHost`",
+  "  to \"\" (the server runs it as a fleet-wide HUNT across all endpoints). Never invent a hostname.",
+  "- Each `vql` MUST be a SINGLE, self-contained, CLIENT-side Velociraptor VQL statement —",
+  "  one `SELECT … FROM <plugin>(…) WHERE …`. glob() uses FORWARD slashes. Do NOT put a blank line",
+  "  inside one query and do NOT make a query only a comment.",
+  "- STRONGLY PREFER raw VQL plugins — they ALWAYS exist: pslist(), netstat(), glob(), stat(),",
+  "  read_file(), hash(), yara(), parse_evtx(), reg_keys(). Network connections → netstat(); processes →",
+  "  pslist(); files → glob(); event logs → parse_evtx(filename='C:/Windows/System32/winevt/Logs/Security.evtx').",
+  "- You MAY reference an `Artifact.<Name>()` ONLY if <Name> appears EXACTLY in the AVAILABLE VELOCIRAPTOR",
+  "  ARTIFACTS list in the user message below. If <Name> is NOT in that list it does NOT exist on this",
+  "  server and the hunt FAILS TO COMPILE (no flow id) — use a raw plugin instead. NEVER invent an",
+  "  artifact name (e.g. Windows.EventLogs.Sysmon / .SecurityLog are NOT universal — check the list).",
+  "- Use each plugin's REAL argument names: parse_evtx() takes `filename=` (a path/glob), NOT `files=`;",
+  "  handles() takes `pid=`, NOT `process=`; read_file() takes `filenames=` (a LIST), NOT `filename=`.",
+  "  Access EVTX fields as `System.EventID.Value` / `EventData.<Name>`. To read one file:",
+  "  `SELECT Data FROM read_file(filenames=['C:/path/file.txt'])`. To read files found by glob,",
+  "  call it inline in the SELECT (no JOIN needed):",
+  "  `SELECT FullPath, read_file(filenames=[FullPath])[0].Data AS Content FROM glob(globs='...')`.",
+  "- VQL has NO SQL `JOIN`. Use inline function calls (e.g. read_file above) or",
+  "  `foreach(row={SELECT … FROM a()}, query={SELECT … FROM b()})` to correlate two plugins.",
+  "- hash() takes ONLY `path=` and returns an object with `.MD5` / `.SHA1` / `.SHA256` — there is NO",
+  "  `hashselect`/`algorithm`/`type` argument (inventing one fails to COMPILE, so the hunt never starts).",
+  "  Get a file's SHA256 as `hash(path=FullPath).SHA256`, computed ONCE as a column, not repeatedly in WHERE:",
+  "  `SELECT FullPath, hash(path=FullPath).SHA256 AS SHA256 FROM glob(globs='…/*.exe') WHERE SHA256 IN (…)`.",
+  "- NEVER hash with a full-disk glob like 'C:/**/*.exe' — it walks every file on the volume and times out.",
+  "  Scope globs to the directories the tradecraft uses (web roots, %TEMP%, Downloads, a service path).",
+  "- For an ABSOLUTE time use timestamp(string='2025-03-14T22:00:00Z'); epoch= takes unix SECONDS (a number).",
+  "- Velociraptor VQL has NO duration-suffix literals. Do NOT write `30d`, `7h`, `2w`. Use seconds",
+  "  arithmetic: `now() - 30 * 86400` (30 days), `now() - 3600` (1 hour). Wrap in `timestamp(epoch=...)`",
+  "  when comparing a timestamp column, e.g. `WHERE Mtime > timestamp(epoch=now() - 30 * 86400)`.",
+  "- Pivot on the case's REAL indicators (the exact hashes, file paths, process/service names, domains,",
+  "  IPs shown) — do NOT invent IOCs the case does not contain.",
+  "- For each: a short `title`; a `rationale` (which task it serves, what the query looks for, how to",
+  "  triage a hit); `severity` (Critical|High|Medium|Low|Info) of the underlying threat; and",
+  "  `mitreTechniques` (relevant technique ids).",
+  "",
+  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
+  JSON.stringify({
+    suggestions: [{
+      taskId: "finding:f3",
+      endpointRelated: true,
+      title: "Enumerate the malicious service on WEB01",
+      rationale: "Task asks to investigate the service-persistence finding on WEB01. Collect the host's services and flag the one whose ImagePath matches the dropped binary; triage by start type and account.",
+      vql: "SELECT Name, DisplayName, PathName, StartMode FROM Artifact.Windows.System.Services() WHERE PathName =~ 'evil\\\\.exe'",
+      targetHost: "WEB01",
+      severity: "High",
+      mitreTechniques: ["T1543.003"],
+    }],
+  }, null, 2),
+].join("\n");
+
+// Translate a free-text analyst request into a runnable hunting query per platform (issue #100).
+// The analyst describes the activity in plain English; the model maps that intent onto each requested
+// platform's REAL schema (the per-platform reference is supplied in the user message). EPHEMERAL like
+// ask()/suggestHunts(): no state change. The Velociraptor query is deployable as-is via launchHunt, so
+// the VQL shape constraints mirror what splitVqlStatements expects (one self-contained statement).
+export const QUERY_TRANSLATE_PROMPT = [
+  "You are a senior DFIR detection engineer. An analyst will give you a request in PLAIN ENGLISH describing",
+  "the activity they want to find in their logs (e.g. \"PowerShell downloading a file and then executing it\",",
+  "\"outbound RDP from this host\", \"new local administrator accounts\"). Translate that intent into a runnable",
+  "hunting query for EACH target platform listed in the user message, grounded in that platform's REAL schema",
+  "(the tables / plugins / field names given for it).",
+  "",
+  "Rules:",
+  "- Emit one entry per TARGET PLATFORM key shown in the user message — use the exact key (velociraptor,",
+  "  defender, elastic, splunk, sigma, yara, suricata). Do NOT invent platforms or emit one not requested.",
+  "- Ground every query in the platform's REAL schema shown for it — its actual tables/plugins/field names.",
+  "  Do NOT invent table or field names; prefer the canonical fields listed.",
+  "- Capture the analyst's FULL intent, including sequencing/relationships where the platform allows it (e.g.",
+  "  a parent→child process relationship, \"download THEN execute\"). When a platform can't express a relation,",
+  "  approximate with the closest field filters and note the limitation in `caveats`.",
+  "- If the request references a specific case entity (\"this host\", \"that IP/hash\"), use the matching value",
+  "  from the PIVOTABLE INDICATORS list — do NOT invent indicators the case does not contain.",
+  "- Velociraptor `query` MUST be a SINGLE, self-contained, CLIENT-side VQL statement — one",
+  "  `SELECT … FROM <plugin>(…) WHERE …`. glob() uses FORWARD slashes; VQL has NO SQL JOIN (use foreach() or",
+  "  inline calls) and NO duration literals (use now() - N * 86400). Do NOT put a blank line inside one query.",
+  "- KQL / ES|QL / SPL queries must be directly runnable (piped where idiomatic). Sigma is a YAML detection",
+  "  rule; Suricata is `alert … (msg:…; …; sid:9000001; rev:1;)`; YARA targets FILE CONTENT.",
+  "- If a platform genuinely cannot express the request (e.g. YARA or Suricata for a pure process-behavior",
+  "  request with no file/network indicator), set `notApplicable` true, leave `query` empty, and explain why",
+  "  in `caveats` — do NOT force a meaningless query.",
+  "- For each entry: a short `label`; the `query`; an `explanation` (how it captures the request + what a hit",
+  "  looks like); optional `caveats` (assumptions, field-mapping notes, what to verify before trusting hits).",
+  "- Also return a one-sentence `interpretation` of how you understood the request, so the analyst can confirm",
+  "  intent before running anything.",
+  "",
+  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
+  JSON.stringify({
+    interpretation: "Find PowerShell processes that download a file and then execute it.",
+    queries: [{
+      platform: "velociraptor",
+      label: "PowerShell download-and-execute (live processes)",
+      query: "SELECT Pid, Ppid, Name, CommandLine, Exe FROM pslist() WHERE Name =~ '(?i)powershell' AND CommandLine =~ '(?i)(DownloadString|DownloadFile|Invoke-WebRequest|iwr|curl|wget)' AND CommandLine =~ '(?i)(Invoke-Expression|iex|Start-Process|-enc)'",
+      explanation: "Lists running PowerShell processes whose command line shows BOTH a download primitive and an execution primitive; a hit is one process doing both.",
+      caveats: "Live process list only — pair with EID 4104 script-block logs / Sysmon 1 for historical coverage.",
+      notApplicable: false,
+    }, {
+      platform: "defender",
+      label: "PowerShell download then execute",
+      query: "DeviceProcessEvents\n| where FileName =~ \"powershell.exe\"\n| where ProcessCommandLine has_any (\"DownloadString\",\"DownloadFile\",\"Invoke-WebRequest\",\"iwr\",\"curl\",\"wget\")\n| where ProcessCommandLine has_any (\"Invoke-Expression\",\"iex\",\"Start-Process\",\"-enc\")",
+      explanation: "Returns PowerShell process events whose command line contains both a download and an execution primitive.",
+      caveats: "Tune the keyword lists; join DeviceNetworkEvents to confirm the download egress.",
+      notApplicable: false,
+    }],
+  }, null, 2),
+].join("\n");
