+++ companion/src/analysis/ai/prompts/investigation.ts
@@ -1,0 +1,249 @@
+/**
+ * Prompts for the analyst's interactive loop: asking a question, reviewing and generating
+ * hypotheses, proposing the next step, and explaining a single event.
+ *
+ * Moved verbatim from pipeline.ts (#384). The text is byte-for-byte what it was: tests/eval/
+ * changeGate.ts hashes these constants, so a reflowed line is indistinguishable from an edited
+ * prompt and would demand a fresh no-regression attestation for a change that is not one.
+ */
+
+
+// Answer a free-form analyst question about ONE case using only its evidence digest.
+export const ASK_PROMPT = [
+  "You are a DFIR analyst assistant answering a SPECIFIC question about ONE investigation, using ONLY the",
+  "case evidence provided below (compromised assets, threat-intel verdicts, attacker path, findings,",
+  "forensic timeline, current questions). Do NOT invent evidence — if the case doesn't show it, say so.",
+  "",
+  "When an ATTACK GRAPH section is present, it lists the case's deterministic CAUSAL relationships —",
+  "process spawns (parent → child), file lineage (wrote → executed), lateral movement (same",
+  "binary/account across hosts), and network connections (source → destination). For multi-hop or",
+  "PATH questions (e.g. 'trace the path from the phishing email to the Domain Controller'), FOLLOW",
+  "these edges end-to-end to reconstruct the route — chain spawn → file → lateral → network hops —",
+  "instead of guessing from the prose timeline alone, and cite the backing [event ids] in",
+  "relatedEventIds. The graph is the ground truth for what led to what.",
+  "",
+  "Pick a status:",
+  "- 'answered': the case evidence clearly settles it. Give the answer and cite the supporting event ids",
+  "  in relatedEventIds.",
+  "- 'partial': suggestive but incomplete evidence. State what is known and what is missing.",
+  "- 'unknown': the case has no evidence either way.",
+  "",
+  "For 'partial' or 'unknown', set 'pointer' to CONCRETE collection guidance — the exact artifact(s) to",
+  "examine or collect and where, named like a DFIR pro would (registry keys, event-log channels, file",
+  "paths, log sources, and the tool / Velociraptor artifact to pull). Examples:",
+  "- USB connected → USBSTOR + MountedDevices + MountPoints2 registry, setupapi.dev.log, and the",
+  "  Microsoft-Windows-DriverFrameworks-UserMode/Operational + Partition/Diagnostic event logs.",
+  "- Data exfiltration → proxy/firewall egress + netflow for large/unusual outbound transfers, cloud-upload",
+  "  logs, DNS logs for tunnelling, EDR network telemetry; look for archive/staging files (.zip/.rar/.7z).",
+  "- Lateral movement → 4624/4672 (logon type 3/10) + 4648, SMB/admin$ access, PsExec/WMI/WinRM artifacts.",
+  "Tailor it to the question and keep 'answer' to a few sentences.",
+  "",
+  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
+  JSON.stringify({
+    answer: "concise answer grounded in the evidence (or what's missing)",
+    status: "answered|partial|unknown",
+    pointer: "which artifact to examine/collect and where (required for partial/unknown)",
+    relatedEventIds: ["e1"],
+  }, null, 2),
+].join("\n");
+
+// On-demand falsification review of the OPEN hypotheses (issue #71) — a focused devil's-advocate pass
+// that explicitly hunts DISCONFIRMING evidence to counter confirmation bias. One text-only call; EPHEMERAL.
+// The recommended status is ADVISORY (the analyst owns the verdict), so the prompt must not claim to change it.
+export const HYPOTHESIS_REVIEW_PROMPT = [
+  "You are a DFIR analyst performing a FALSIFICATION REVIEW of the OPEN investigative hypotheses for ONE",
+  "case, using ONLY the evidence provided below (compromised assets, threat-intel verdicts, attacker path,",
+  "findings, forensic timeline). Your job is to fight CONFIRMATION BIAS: for each hypothesis, weigh the",
+  "evidence for AND — most importantly — actively look for evidence AGAINST it.",
+  "",
+  "For EACH hypothesis listed under OPEN HYPOTHESES TO REVIEW (reference it by its EXACT id):",
+  "- supportingEvidence: plain-English bullets for the evidence that SUPPORTS the hypothesis (or [] if none).",
+  "- refutingEvidence: plain-English bullets for the evidence that REFUTES or WEAKENS it — the disconfirming",
+  "  lens. Note absent evidence you WOULD expect if the hypothesis were true (e.g. 'no phishing email in the",
+  "  mail logs despite an inbox-rule change'). Return [] only if you genuinely find nothing against it.",
+  "- recommendedStatus: your ADVISORY verdict — 'supported' (evidence clearly confirms), 'refuted' (evidence",
+  "  clearly contradicts), or 'unknown' (inconclusive / needs collection). This is a RECOMMENDATION for the",
+  "  analyst; you are NOT changing the status. Use 'open' only if it should keep being actively tested as-is.",
+  "- rationale: one short paragraph justifying the recommendation, weighing support against refutation.",
+  "- relatedEventIds: the EXACT ids of the case events your bullets cite. Never invent an id.",
+  "",
+  "Ground everything ONLY in the supplied evidence — do NOT invent events, hosts, or files. Judge each",
+  "hypothesis by how well it survives disconfirming evidence, not by how much supports it.",
+  "",
+  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
+  JSON.stringify({
+    reviews: [{
+      hypothesisId: "<the exact id shown>",
+      title: "<the hypothesis title>",
+      supportingEvidence: ["evidence for it"],
+      refutingEvidence: ["evidence against it, incl. expected-but-absent evidence"],
+      recommendedStatus: "supported | refuted | unknown | open",
+      rationale: "why, weighing support vs refutation",
+      relatedEventIds: ["<event id>"],
+    }],
+  }, null, 2),
+].join("\n");
+
+// Hypothesise attacker actions for TIMELINE GAPS (issue #96). The deterministic gap detector has
+// already flagged suspiciously silent periods; the model reads each gap's bounding context — the
+// events just BEFORE the silence and just AFTER — and infers what the attacker most likely did during
+// the hole (e.g. cleared the log to hide credential dumping, disabled EDR before lateral movement).
+// It is grounded ONLY in the surrounding events; it does NOT invent activity. It also names which
+// SHADOW ARTIFACTS (from the catalog ids in the user message) would best reconstruct each window —
+// the deterministic catalog supplies the actual collection VQL, so the model only ranks relevance.
+export const GAP_HYPOTHESIS_PROMPT = [
+  "You are a senior DFIR analyst reasoning about COVERAGE GAPS in ONE investigation's forensic timeline.",
+  "Each gap below is a stretch where logging went silent — a COMPLETE gap (every source dark) is the",
+  "classic signature of cleared Windows Event Logs, a stopped collector/auditd, or disabled EDR. For",
+  "EACH gap, hypothesise what the attacker most likely did DURING the silence, reasoning from the events",
+  "immediately BEFORE the gap (what they were doing) and immediately AFTER (the state when logging",
+  "resumed).",
+  "",
+  "Rules:",
+  "- Ground every hypothesis ONLY in the surrounding events shown. Do NOT invent specific hosts, files,",
+  "  or accounts the context does not mention. If the surrounding events are too sparse to say anything,",
+  "  give a low confidence and say the gap is unexplained.",
+  "- Prefer the explanation that fits the tradecraft: a complete silence right after initial access often",
+  "  hides discovery/credential-access/defense-evasion (clearing logs to cover the next step); a gap",
+  "  bracketed by a logon and later persistence often hides lateral movement or staging.",
+  "- `gapId` MUST be the exact [gap-N] id shown for the gap the hypothesis is about. Emit at most one",
+  "  hypothesis per gap. It is fine to skip a gap that is plainly benign (e.g. an expected overnight quiet).",
+  "- `hypothesis`: 2-4 sentences naming the most probable attacker activity and WHY it fits the context.",
+  "- `attackerActions`: a few concrete candidate actions that would produce this exact gap.",
+  "- `confidence`: 0-100, honest — sparse context or an equally-likely benign explanation means LOW.",
+  "- `severity`: Critical|High|Medium|Low|Info — how serious the hypothesised activity would be.",
+  "- `mitreTechniques`: ATT&CK ids for the hypothesised actions (e.g. T1070.001 for cleared event logs).",
+  "- `recommendedArtifactIds`: from the SHADOW ARTIFACTS list in the user message, the ids whose data",
+  "  would best confirm THIS hypothesis (e.g. prefetch/amcache/shimcache/bam for execution, usn-journal/",
+  "  mft/lnk-files for file activity, srum for exfiltration). Use ONLY ids from that list.",
+  "",
+  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
+  JSON.stringify({
+    hypotheses: [{
+      gapId: "gap-1",
+      hypothesis: "The Security log was cleared immediately after the initial RDP logon and before the service-install seen on resume, consistent with the attacker wiping logs to hide credential access and staging during the silence.",
+      attackerActions: ["Cleared the Windows Security event log (wevtutil cl / EventLog API)", "Dumped LSASS or ran a discovery tool while logging was off"],
+      confidence: 55,
+      severity: "High",
+      mitreTechniques: ["T1070.001", "T1003.001"],
+      recommendedArtifactIds: ["prefetch", "amcache", "usn-journal", "srum"],
+    }],
+  }, null, 2),
+].join("\n");
+
+// Memory-forensics "Next-Step" agent (issue #101). The case already has Volatility 3 / Rekall output
+// imported as forensic events (the process tree, network connections, malfind injected code, command
+// lines, services, modules). Read that memory evidence, identify the ANOMALIES, and propose the EXACT
+// next Volatility 3 command the analyst should run to dig deeper. The agent CONSUMES the enumeration
+// (it does not re-implement Volatility) — it reasons over the rows and recommends the next plugin.
+export const MEMORY_NEXTSTEP_PROMPT = [
+  "You are a senior memory-forensics analyst guiding an ITERATIVE Volatility 3 investigation. Below is",
+  "the memory evidence ALREADY imported from a RAM image (Volatility 3 / Rekall output): the process",
+  "tree (process name, PID, PPID, parent name, start time, command line), network connections, malfind",
+  "(executable/injected private memory), command lines, services, and loaded modules. Identify the",
+  "ANOMALIES and, for each, propose the EXACT next Volatility 3 command the analyst should run to dig in.",
+  "",
+  "What counts as an anomaly (reason from the evidence shown, do NOT invent processes/PIDs):",
+  "- Process-tree masquerading / wrong parentage: svchost.exe NOT parented by services.exe; lsass.exe,",
+  "  csrss.exe, services.exe, wininit.exe with the wrong/absent parent; an unparented process; a system",
+  "  binary running from a non-system path; a user app spawning cmd.exe/powershell.exe.",
+  "- Injected/executable private memory (malfind hits) → confirm whether it is real injection or benign.",
+  "- Suspicious or external network connections owned by an unexpected process (possible C2/beacon).",
+  "- LOLBin / encoded-PowerShell / unusual command lines.",
+  "- A persistence-looking service or an unsigned/odd module.",
+  "",
+  "APPLY FALSE-POSITIVE AWARENESS (this is the most important part — every malfind hit is ingested as",
+  "High, but many RWX/executable-private-memory regions are BENIGN). Before proposing a dig-in step, ask",
+  "whether the hit is expected for that process:",
+  "- Security/AV engines legitimately use RWX: MsMpEng.exe (Microsoft Defender), MpDefenderCoreService,",
+  "  and third-party AV/EDR. These are the #1 malfind false positive.",
+  "- .NET/CLR and other JIT compilers emit RWX: powershell.exe, processes hosting the CLR, and",
+  "  JavaScript/Java/Lua JITs (browsers — chrome/msedge/firefox, node, java). RWX here is normal JIT.",
+  "- Some legitimate packers/installers and SearchHost.exe/Search/UI shell processes also show RWX.",
+  "When the malfind hit is on such a process AND nothing else about it is anomalous (correct image path,",
+  "correct parent, no suspicious cmdline/connection), SAY SO in the `anomaly`/`rationale`, set `severity`",
+  "to Low or Info, and make the next step a quick LEGITIMACY CONFIRMATION — `windows.cmdline` /",
+  "`windows.dlllist` to verify the image path, signer, and loaded modules — rather than dumping every",
+  "region. Reserve High/Critical and a real dig-in for genuinely unexpected processes, wrong parentage,",
+  "bad paths, or malfind correlated with a suspicious connection/command line.",
+  "",
+  "Rules:",
+  "- Each `command` MUST be a single, real, copy-pasteable Volatility 3 command. Use `vol -f <image>`",
+  "  as the prefix (the analyst substitutes their image path for <image>) followed by a REAL Volatility 3",
+  "  plugin and its REAL options, e.g.:",
+  "    vol -f <image> windows.malfind --pid 1234",
+  "    vol -f <image> windows.dlllist --pid 1234",
+  "    vol -f <image> windows.cmdline --pid 1234",
+  "    vol -f <image> windows.handles --pid 1234",
+  "    vol -f <image> windows.netscan",
+  "    vol -f <image> windows.pstree",
+  "    vol -f <image> windows.getsids --pid 1234",
+  "    vol -f <image> windows.svcscan",
+  "  Use Linux/Mac plugin names (linux.* / mac.*) instead if the evidence is clearly from that OS.",
+  "  Use the REAL plugin/option names — do NOT invent plugins or flags, and do NOT use Volatility 2",
+  "  syntax (no `--profile`, no `vol.py -f mem.raw pslist`-style v2 plugin names).",
+  "- STRONGLY PREFER commands that produce a TABLE the analyst can paste/import straight back into this",
+  "  tool (malfind, cmdline, handles, dlllist, netscan, pstree, svcscan, getsids, privileges, …). Do NOT",
+  "  add `--dump` and do NOT suggest a plain `windows.dumpfiles`/`windows.procdump` as the step UNLESS",
+  "  dumping is genuinely the right next move — a dump writes a RAW BINARY .dmp/.exe to disk, which is",
+  "  NOT something this tool can ingest. When you DO recommend a dump, the `rationale` MUST say the .dmp",
+  "  is for OFFLINE analysis (YARA/`capa`/`strings`/upload to a malware sandbox) and that the analyst",
+  "  imports THOSE results back (this tool ingests sandbox reports) — the .dmp itself is not re-imported.",
+  "- PREFER suggesting plugins that have NOT been run yet (the user message lists the already-imported",
+  "  plugins) when they would advance the investigation — the point is the NEXT step, not re-running",
+  "  what is already on the timeline. Pivot on a SPECIFIC PID/process from the evidence wherever the",
+  "  plugin takes `--pid`; set `pid` to that PID.",
+  "- Prefer a few HIGH-SIGNAL next steps over many near-duplicates. If nothing in the evidence looks",
+  "  anomalous, it is fine to return fewer (or no) suggestions.",
+  "- For each: a short `anomaly` (the observation that triggered it, naming the real process/PID); the",
+  "  `command`; the `plugin` it runs (e.g. windows.malfind); a `rationale` (why run it + how to triage",
+  "  what it returns); `severity` (Critical|High|Medium|Low|Info) of the underlying anomaly;",
+  "  `pid` (the targeted PID, or \"\"); and `mitreTechniques` (relevant ATT&CK ids, e.g. T1055 for injection).",
+  "",
+  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
+  JSON.stringify({
+    suggestions: [{
+      anomaly: "svchost.exe (PID 1234) is parented by explorer.exe (PID 4500), not services.exe — classic masquerading.",
+      command: "vol -f <image> windows.malfind --pid 1234",
+      plugin: "windows.malfind",
+      rationale: "A mis-parented svchost is a strong injection/masquerade signal. malfind dumps executable private memory in the process; triage any MZ/shellcode region by yara-scanning the dump and pivot on its imports.",
+      severity: "High",
+      pid: "1234",
+      mitreTechniques: ["T1055", "T1036.005"],
+    }],
+  }, null, 2),
+].join("\n");
+
+// Explain a SINGLE forensic event in context — what happened, why it matters, ATT&CK mapping,
+// pivot queries, and evidence for/against maliciousness (issue #141). EPHEMERAL (no state change).
+export const EXPLAIN_EVENT_PROMPT = [
+  "You are a DFIR (Digital Forensics & Incident Response) analyst explaining ONE specific forensic",
+  "event to another analyst. Using ONLY the case evidence provided (compromised assets, threat-intel",
+  "verdicts, findings, nearby timeline events), explain:",
+  "",
+  "- WHAT happened: describe the event in plain English",
+  "- WHY it matters: its significance to this specific investigation",
+  "- NORMAL vs. SUSPICIOUS: would this event be expected behavior, or is it clearly attacker activity?",
+  "- ATTACK MAPPING: what the tagged ATT&CK technique(s) mean in this context (or empty if none tagged)",
+  "- PIVOT QUERIES: 1–3 concrete follow-up hunts (Velociraptor VQL, Defender/Sentinel KQL, or Splunk",
+  "  SPL) that would collect corroborating or contradicting evidence for this specific event",
+  "- EVIDENCE FOR: what in the case makes this event look malicious",
+  "- EVIDENCE AGAINST: any plausible benign explanation (be honest; do not dismiss ambiguity)",
+  "",
+  "Ground every claim in the provided case context. If context is insufficient, say so explicitly.",
+  "Pivot queries must use real field names for the platform; make them runnable as-is or with minimal",
+  "schema edits. Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
+  JSON.stringify({
+    summary: "what happened, in plain English",
+    whyItMatters: "why this event matters to THIS investigation (1–2 sentences)",
+    normalContext: "is this kind of event normal in non-incident environments?",
+    suspiciousIndicators: "what specifically makes this instance suspicious",
+    attackMapping: "ATT&CK technique(s) and what they mean in context (empty string if none tagged)",
+    pivotQueries: [
+      { platform: "velociraptor|kql|spl|other", query: "the runnable query", rationale: "what it would prove/disprove" },
+    ],
+    evidenceFor: "case evidence supporting malicious interpretation",
+    evidenceAgainst: "plausible benign explanation (or empty string if clearly malicious)",
+    relatedEventIds: ["event ids from the context that support the explanation"],
+  }, null, 2),
+].join("\n");
