+++ companion/src/analysis/pipeline.ts
@@ -1,4 +1,4 @@
-import { readFileSync, createReadStream } from "node:fs";
+import { createReadStream } from "node:fs";
 import { mkdir, writeFile } from "node:fs/promises";
 import { createInterface } from "node:readline";
 import { join as joinPath } from "node:path";
@@ -27,7 +27,6 @@ import type { StateLock } from "./stateLock.js";
 import { sortByEventTime } from "./forensicSort.js";
 import { segmentSessions, sessionEnvOptions } from "./sessionSegmentation.js";
 import { applySeverityFloor } from "./severityFloor.js";
-import { EXAMPLE_IMPORTER_SPEC } from "./importerSpec.js";
 import type { ExternalImporter } from "./declarativeImporter.js";
 import { parseJsonLoose } from "./extractJson.js";
 import { applyFalsePositive, buildFalsePositiveContext, buildAuthorizedContextBlock, filterFalsePositiveEvents, type FalsePositiveStore } from "./falsePositive.js";
@@ -61,10 +60,20 @@ import { loadKnownPlaybooks } from "./knownPlaybooksData.js";
 import { buildPlaybookMatchResult, playbookMatchEnvOptions } from "./playbookMatch.js";
 import { buildSynthesisCoverage, type SynthMetaStore, type SynthesisCoverage } from "./synthMeta.js";
 import { AiCostStore, bucketForLabel } from "./aiCost.js";
+// The prompt registry moved to ai/prompts/ (#384). Imported for the pipeline's own use and
+// re-exported below, because 23 modules and the eval harness import these names from here.
+import {
+  getAskPrompt, getCsvPrompt, getExecSummaryPrompt, getExplainEventPrompt, getFpSimilarityPrompt,
+  getGapHypothesisPrompt, getHuntSuggestPrompt, getHypothesisReviewPrompt,
+  getLogPrompt, getMemoryNextStepPrompt, getNarrativePrompt, getObservePrompt, getPlaybookHuntPrompt,
+  getQueryTranslatePrompt, getReconcilePrompt, getRemediationPrompt, getSessionSummaryPrompt,
+  getStarredReportPrompt, getSynthesisPrompt, getSystemPrompt, getTaggerRulePrompt,
+  getViewSummaryPrompt,
+} from "./ai/prompts/index.js";
+export * from "./ai/prompts/index.js";
 import { CorrelationProfileStore } from "./correlationProfile.js";
 import type { SecondOpinionStore } from "./secondOpinionStore.js";
 import {
-  RECONCILE_PROMPT,
   buildSecondOpinion,
   buildReconcilePrompt,
   reconcileResponseSchema,
@@ -233,1219 +242,10 @@ async function dumpRedactedImage(
   }
 }
 
-export const SYSTEM_PROMPT = [
-  "You are a DFIR analyst assistant. You are shown screenshots from a forensic investigation. The",
-  "evidence may come from ANY security tool, including: Velociraptor, EDR/XDR consoles (CrowdStrike",
-  "Falcon, Microsoft Defender for Endpoint, SentinelOne, Carbon Black), SIEM/log UIs (Splunk, Elastic/",
-  "Kibana), VirusTotal, and raw Windows artifacts. Plus a summary of findings already recorded.",
-  "Update existing findings by their id; never create a duplicate finding for a topic already",
-  "listed. Open a thread for any lead you start chasing and close it by id when resolved.",
-  "",
-  "CRITICAL — FORENSIC TIMELINE: forensic artifacts on screen carry REAL timestamps (process",
-  "create time, file MAC times, logon time, prefetch run time, scheduled-task time, registry",
-  "write time, network connection time, etc.). For every dated incident event you can read, emit",
-  "a forensicEvents entry whose timestamp is read FROM THAT ROW's OWN time column in the image",
-  "(e.g. the 'Timestamp'/'EventTime' column of the results table). EMIT EVERY TIMESTAMP IN UTC —",
-  "ISO-8601 with a trailing 'Z'. If the artifact shows a timezone (an offset like +02:00, or an",
-  "abbreviation like EST/CEST/UTC+2), CONVERT the time to UTC. If it shows NO timezone, keep the",
-  "wall-clock time and just add a trailing 'Z' (forensic tooling overwhelmingly displays UTC) —",
-  "never shift a timezone-less time.",
-  "NEVER use the screenshot capture time or the current time. If a row has no visible event time,",
-  "set its timestamp to an empty string \"\" — do NOT substitute the capture/current time.",
-  "These reconstruct WHEN the attack happened on the SYSTEM(S) UNDER INVESTIGATION.",
-  "",
-  "YOUR PRIMARY JOB — EXTRACT THE ARTIFACT ROWS. Velociraptor/EDR/SIEM tables ARE the evidence.",
-  "Extract EVERY data row that represents real host/attacker activity as a forensicEvent, using the",
-  "timestamp in that row. This is the most important output — do NOT be shy. High-value rows you must",
-  "ALWAYS capture: antivirus / EDR alerts (Microsoft Defender, Sysmon), Security event-log logons,",
-  "process executions, network connections, scheduled tasks, services, registry changes, file MAC",
-  "times. Example — a Defender 'Operational' row:",
-  "  Timestamp 2026-05-26T12:25:36Z | Defender Alert (Severe) | VirTool:Win32/Kekeo.A!MTB |",
-  "  Path C:\\Users\\srv\\Downloads\\Rubeus.exe | User ADATUMLAB\\srv",
-  "  → forensicEvent: timestamp '2026-05-26T12:25:36Z', severity 'Critical', description",
-  "    'Microsoft Defender flagged VirTool:Win32/Kekeo (Rubeus.exe) for user ADATUMLAB\\\\srv at",
-  "    C:\\\\Users\\\\srv\\\\Downloads\\\\Rubeus.exe'. Add IOCs (Rubeus.exe, the threat name) and a",
-  "    finding (credential-theft tool / Kekeo). NEVER drop a malware/threat detection.",
-  "",
-  "EDR/XDR & SIEM DETECTIONS ARE EVIDENCE — extract EVERY one. A CrowdStrike Falcon / Defender for",
-  "Endpoint / SentinelOne 'detections' view, OR a SIEM alerts view (Splunk notable events / Enterprise",
-  "Security, Elastic Security alerts, Microsoft Sentinel incidents, QRadar offenses, a Kibana/Grafana",
-  "panel listing security alerts), is a CONSOLE OF FINDINGS — NOT a 'dashboard' to skip. Each detection",
-  "/ alert / notable / offense — with its severity, rule/detection name, triggering process / command",
-  "line / query, file path, hash, host, user, source/dest IP, and MITRE tactic & technique — IS a real",
-  "event: emit it as a forensicEvent AND raise a finding. Use the alert's OWN time as the event time.",
-  "Example — a CrowdStrike High detection:",
-  "  Jun 1 2026 12:37:35 | High | ShadowMark.exe on ALCLIENT04 by ADATUMLAB\\Srv |",
-  "  cmdline '\\\"C:\\\\Users\\\\srv\\\\Desktop\\\\New folder\\\\ShadowMark.exe\\\" /action:add /target:sac1$' |",
-  "  SHA256 2eeba4c80a6f91f06784c0c699512c22ff132233c71af336a423414cc84f574a | AI Powered IOA / Malicious File",
-  "  → forensicEvent severity 'High', description 'CrowdStrike flagged ShadowMark.exe (High, Malicious",
-  "    File / AI IOA) run by ADATUMLAB\\\\Srv on ALCLIENT04: cmdline … /action:add /target:sac1$; parent",
-  "    process killed'; IOCs ShadowMark.exe + the SHA256; a finding (suspected machine-account/SAC abuse).",
-  "Example — a Splunk notable / Elastic Security alert:",
-  "  2026-06-01T09:14:22Z | High | rule 'Brute Force - Multiple Failed Logons' | host ALCLIENT07 |",
-  "  user ADATUMLAB\\\\jdoe | src 10.0.0.5 | T1110",
-  "  → forensicEvent severity 'High', description 'SIEM rule \\\"Brute Force - Multiple Failed Logons\\\"",
-  "    fired for ADATUMLAB\\\\jdoe on ALCLIENT07 from 10.0.0.5'; IOC 10.0.0.5; a finding (brute force,",
-  "    T1110). Multiple near-identical detections/alerts collapse into ONE finding, but keep each",
-  "    distinct process/host/user/rule as its own event.",
-  "",
-  "SEVERITY/LEVEL COLUMN — STRONG FINDING SIGNAL. If a row has its own Severity / Level / Criticality",
-  "/ Risk / Confidence column and it reads Critical, High, or Severe (or a high numeric score), treat",
-  "it as important by default (~90% of the time it IS a finding): (a) ALWAYS emit it as a forensicEvent",
-  "with that severity mapped to ours (Severe→Critical, High→High), and (b) ALSO raise a finding for it",
-  "in this response — do not wait for the synthesis pass. Only skip the finding if the row is clearly",
-  "benign/expected (e.g. an informational rule the client confirmed legitimate). Never silently drop a",
-  "Critical/High row.",
-  "",
-  "DESCRIBE EACH EVENT BY WHAT HAPPENED ON THE SYSTEM, not by the tool you saw it in. Write the",
-  "artifact's own facts. WRONG: 'Velociraptor EventLog shows a Defender alert…'. RIGHT: 'Microsoft",
-  "Defender flagged Rubeus.exe…'. Do NOT start an event with 'Velociraptor', 'VolWeb', 'the dashboard",
-  "shows', or append 'observed in <tool>' — name the host, process, user, IP, etc. instead.",
-  "",
-  "DO NOT ADD NON-EVENTS. Two things are NOT forensic events (skip them — but never let this cause you",
-  "to drop a real artifact row or a detection): (1) operating the tooling / your own workflow — running",
-  "a hunt or query, 'EventLog analysis performed', 'Response and Monitoring accessed', 'data",
-  "collection', 'analysis completed', clicking/scrolling; (2) the bare act of opening a tool with NO",
-  "data on screen — 'Access to VolWeb', 'DFIR Companion dashboard access observed', an empty/loading",
-  "panel or login page. This is NARROW: a detections/alerts console FULL of detections or alerts",
-  "(CrowdStrike, Defender, SentinelOne, Splunk, Elastic, Sentinel, QRadar, Kibana/Grafana security",
-  "panels) is EVIDENCE — extract them, do not dismiss the page as 'navigation'. If a screen names a",
-  "process, file, command line, query, hash, host, user, IP, rule/alert name, or threat, it has",
-  "content → extract it. NEVER stamp an event with the screenshot capture time — if a row truly has no",
-  "time column, use \"\".",
-  "",
-  "AFFECTED ASSET: for each forensicEvent, set 'asset' to the host/computer/FQDN the event happened",
-  "ON — read it from the row's Computer/Hostname/FQDN/Endpoint/Device column, the console's host",
-  "field, or the window title (e.g. 'ALCLIENT07', 'dc01.adatumlab.local'). Leave 'asset' as \"\" only",
-  "if no host is visible. This ties each indicator to the machine it was seen on.",
-  "",
-  "ATTACKER PATH: in 'attackerPath', narrate the adversary's progression in kill-chain order",
-  "(initial access → execution → persistence → priv-esc → lateral movement → C2 → exfil/impact),",
-  "citing finding ids and event times. Refine it as new evidence arrives.",
-  "",
-  "FINDINGS CONFIDENCE: every finding MUST include a 'confidence' field (integer 0–100) — your",
-  "certainty that this is real attacker activity, not a false positive. 95+ = confirmed hit; 70–90 =",
-  "strongly suspicious; 40–69 = plausible; <40 = speculative. Do NOT omit this field. Also include a",
-  "'confidenceReason' field — one short sentence on why (e.g. evidence strength, whether other tools",
-  "corroborate it, or your own certainty).",
-  "",
-  "Return ONLY raw JSON (no markdown code fences, no prose) with EXACTLY this shape — every",
-  "finding/ioc/technique/thread/event MUST be an OBJECT with these keys, never a bare string:",
-  "",
-  JSON.stringify(
-    {
-      findings: [
-        {
-          id: "f1",
-          severity: "Critical|High|Medium|Low|Info",
-          confidence: 85,
-          confidenceReason: "why this score",
-          title: "short title",
-          description: "what was observed and why it matters",
-          relatedIocs: ["i1"],
-          mitreTechniques: ["T1059"],
-          status: "open|confirmed|dismissed",
-        },
-      ],
-      iocs: [{ id: "i1", type: "ip|domain|hash|file|process|url|other", value: "the indicator" }],
-      mitreTechniques: [{ id: "T1059", name: "Command and Scripting Interpreter" }],
-      forensicEvents: [
-        {
-          id: "e1",
-          timestamp: "2026-05-20T14:03:00Z",
-          description: "powershell.exe spawned encoded command (from prefetch run time)",
-          severity: "Critical|High|Medium|Low|Info",
-          mitreTechniques: ["T1059.001"],
-          asset: "ALCLIENT07",
-        },
-      ],
-      threadsOpened: [{ id: "t1", description: "lead being chased" }],
-      threadsClosed: ["t0"],
-      timelineNote: "one sentence on what you reviewed in this batch of screenshots",
-      attackerPath: "kill-chain narrative of how the attacker progressed, citing finding ids and times",
-      summary: "running executive summary of the whole investigation so far",
-    },
-    null,
-    2,
-  ),
-  "",
-  "confidence is 0–100 (your certainty that the finding represents real attacker activity, not a false positive).",
-  "If a section has nothing new, return it as an empty array (or empty string for text fields).",
-].join("\n");
 
-// Extraction prompt for an imported CSV (a Velociraptor/EDR result export). Like
-// SYSTEM_PROMPT but the evidence is structured ROWS, not screenshots: each row is a
-// forensic record and we want its real timestamp read from the row's own time column.
-export const CSV_SYSTEM_PROMPT = [
-  "You are a DFIR analyst assistant. You are given ROWS from a CSV export of forensic results",
-  "(typically a Velociraptor artifact or EDR query: process listings, Windows event-log rows,",
-  "netstat, prefetch, $MFT, scheduled tasks, services, shellbags, AmCache, UserAssist, etc.),",
-  "plus a summary of findings already recorded.",
-  "",
-  "Each data row IS the evidence. For EVERY row that represents real host/attacker activity, emit a",
-  "forensicEvents entry whose 'timestamp' is read FROM THAT ROW's OWN time column — pick the most",
-  "relevant time field present (e.g. Mtime/Btime/Ctime/Atime, Timestamp, EventTime, Created /",
-  "CreationTime, StartTime, RunTime, LastRun, FirstSeen, _ts). EMIT IT IN UTC — ISO-8601 with a",
-  "trailing 'Z': convert any shown timezone offset to UTC; if the column carries no timezone, keep",
-  "the wall-clock time and add 'Z' (these exports are UTC by default) — never shift a naive time.",
-  "If a row has NO usable event time, set timestamp to \"\" — NEVER substitute the current time.",
-  "Give each event a severity and map it to MITRE technique ids where clear.",
-  "",
-  "SEVERITY/LEVEL COLUMN — STRONG FINDING SIGNAL. If a row has its own Severity / Level / Criticality",
-  "/ Risk column reading Critical, High, or Severe (or a high numeric score), treat it as important by",
-  "default (~90% of the time it IS a finding): ALWAYS emit the forensicEvent with that severity",
-  "(Severe→Critical), AND also raise a finding for it in this response. Never silently drop a",
-  "Critical/High row — only skip the finding if the row is clearly benign/expected.",
-  "",
-  "Also surface concrete IOCs present in the rows (ips, domains, hashes, malicious file/process",
-  "names, URLs). Set timelineNote to one short sentence naming the artifact and the columns you read.",
-  "",
-  "AFFECTED ASSET: set each event's 'asset' to the host/computer/FQDN from the row's Computer/Hostname/",
-  "Fqdn/Endpoint/Device column (or the export's host); leave \"\" if none — it ties each indicator to its host.",
-  "",
-  "FINDINGS CONFIDENCE: every finding MUST include a 'confidence' field (integer 0–100) — your",
-  "certainty that this is real attacker activity, not a false positive. Do NOT omit this field. Also",
-  "include a 'confidenceReason' field — one short sentence on why.",
-  "",
-  "Return ONLY raw JSON (no markdown fences). Every event/ioc MUST be an OBJECT. Shape:",
-  "",
-  JSON.stringify(
-    {
-      findings: [
-        { id: "f1", severity: "Critical|High|Medium|Low|Info", confidence: 90, confidenceReason: "why this score", title: "short title (raise for any Critical/High row)", description: "what was detected and why it matters", relatedIocs: ["i1"], mitreTechniques: ["T1059"], status: "open" },
-      ],
-      iocs: [{ id: "i1", type: "ip|domain|hash|file|process|url|other", value: "the indicator" }],
-      mitreTechniques: [{ id: "T1059", name: "Command and Scripting Interpreter" }],
-      forensicEvents: [
-        { id: "e1", timestamp: "2026-05-20T14:03:00Z", description: "what happened (cite the row's key columns)", severity: "Critical|High|Medium|Low|Info", mitreTechniques: ["T1059.001"], asset: "ALCLIENT07" },
-      ],
-      threadsOpened: [],
-      threadsClosed: [],
-      timelineNote: "read N rows of <artifact> (time column: <col>)",
-      attackerPath: "",
-      summary: "",
-    },
-    null,
-    2,
-  ),
-].join("\n");
 
-// Generic log-line extraction (firewall, syslog, sshd/auth.log, IIS/nginx/Apache
-// access, Windows event-log .txt exports, application logs). Each LINE is the
-// evidence — the model picks out whichever timestamp format the source uses
-// (RFC 3164 syslog "May 28 09:00:01", ISO-8601, IIS "yyyy-MM-dd HH:mm:ss",
-// Apache "[28/May/2026:09:00:01 +0000]", epoch seconds, etc.).
-export const LOG_SYSTEM_PROMPT = [
-  "You are a DFIR analyst assistant triaging a log file uploaded as evidence (typical sources:",
-  "firewall logs — Cisco ASA, pfSense, iptables, Palo Alto, Fortinet; syslog; Windows event-log",
-  "text exports; sshd / auth.log; Apache/IIS/nginx access logs; VPN/IKE; application logs).",
-  "",
-  "The raw log has already been DEDUPLICATED for you: identical recurring lines are collapsed into",
-  "PATTERNS. Each pattern below shows ×<count> (how many times it occurred), the first and last time",
-  "it was seen, and one example line. Treat each pattern as ONE candidate event representing all of",
-  "its occurrences — do NOT emit one event per occurrence.",
-  "",
-  "BE SELECTIVE. A forensic timeline is for SECURITY-RELEVANT activity, not routine operations.",
-  "Emit a forensicEvents entry ONLY for patterns that are suspicious, anomalous, or investigation-",
-  "worthy, e.g.: authentication failures / brute force, blocked or denied traffic from unusual",
-  "sources, port/host scans, IDS/IPS or AV hits, privilege changes, account or config changes,",
-  "unexpected outbound connections / beaconing / data transfer, malware indicators, or an abnormal",
-  "VOLUME of failures suggesting brute force or DoS.",
-  "",
-  "SKIP routine operational noise — do NOT emit events for it: normal VPN/IPsec rekeying, IKE",
-  "keying attempts and retransmission/timeout chatter, heartbeats, successful benign connections,",
-  "and informational/debug lines. A high ×count alone does NOT make a pattern suspicious; benign",
-  "infrastructure noise (e.g. a tunnel repeatedly re-keying) should be skipped even at high volume.",
-  "This applies to sudo/privilege lines too, not just network chatter: a NAMED account running a",
-  "SPECIFIC, scoped sudo command (service restart/status, package install/upgrade, listing firewall",
-  "rules, reading a config) is routine sysadmin/SRE work — grade it Info/Low, NOT High, even if it",
-  "recurs often or several different accounts each do their own such commands across the observation",
-  "window (many people doing their normal jobs is not a 'privilege escalation campaign'). Reserve",
-  "High/Critical sudo severity for a genuine anomaly: a brand-new/never-seen account gaining sudo, a",
-  "user or group ADDED to sudoers/wheel/admin, an interactive root shell (sudo -i / su -) opened by",
-  "an account that doesn't normally have one, or sudo usage immediately followed by credential-",
-  "dumping/exfil/tampering commands. Likewise, auth FAILURES are only brute-force-worthy when there's",
-  "a clear escalating pattern (many failures in a short window from ONE source against one/few",
-  "accounts, ideally followed by a success) — a handful of scattered failures spread across many",
-  "different users/hosts over hours is ordinary human error (mistyped/expired passwords), not an",
-  "attack indicator, and should stay Info or be skipped entirely.",
-  "If NOTHING in this batch is security-relevant, return an empty forensicEvents array — that is the",
-  "correct, expected answer for a clean/noisy operational log.",
-  "",
-  "For each event you DO emit, AGGREGATE the whole pattern into one entry:",
-  "  - 'timestamp'    = the pattern's FIRST occurrence time, IN UTC — ISO-8601 with a trailing 'Z':",
-  "                     convert a shown offset (e.g. '+0000', '+02:00') to UTC; a timezone-less syslog",
-  "                     time like 'May 28 09:00:01' keeps its wall-clock and just gets a 'Z'; if there",
-  "                     is no usable time set it to \"\" — NEVER use the current time,",
-  "  - 'endTimestamp' = the pattern's LAST occurrence time, in UTC (omit if same as first / single event),",
-  "  - 'count'        = the pattern's ×<count> occurrence number (copy it verbatim),",
-  "  - 'description'  = an aggregate summary that STATES THE COUNT and time span, e.g.",
-  "                     '20 failed SSH logins for root from 1.2.3.4 between 09:00:01 and 09:04:12'.",
-  "Give each event a severity and map it to MITRE technique ids where clear.",
-  "",
-  "Also surface concrete IOCs present in the suspicious patterns (source/destination ips, domains,",
-  "hashes, URLs, suspicious user/process/file names). Do NOT invent findings or an attacker path —",
-  "those come from a later holistic synthesis pass. Set timelineNote to one short sentence naming",
-  "the log source you inferred (e.g. 'pfSense filter log', 'sshd auth.log', 'strongSwan IKE log').",
-  "",
-  "Return ONLY raw JSON (no markdown fences). Every event/ioc MUST be an OBJECT. Shape:",
-  "",
-  JSON.stringify(
-    {
-      findings: [],
-      iocs: [{ id: "i1", type: "ip|domain|hash|file|process|url|other", value: "the indicator" }],
-      mitreTechniques: [{ id: "T1110", name: "Brute Force" }],
-      forensicEvents: [
-        { id: "e1", timestamp: "2026-05-20T14:03:00Z", endTimestamp: "2026-05-20T14:07:55Z", count: 20, description: "20 failed logins for 'admin' from 1.2.3.4 (possible brute force)", severity: "Critical|High|Medium|Low|Info", mitreTechniques: ["T1110.001"] },
-      ],
-      threadsOpened: [],
-      threadsClosed: [],
-      timelineNote: "triaged N pattern(s) of <inferred log source>",
-      attackerPath: "",
-      summary: "",
-    },
-    null,
-    2,
-  ),
-].join("\n");
 
-// Holistic synthesis: turn the accumulated forensic timeline into analytic
-// conclusions (findings, MITRE, attacker path). Findings/attacker-path need the
-// WHOLE picture, which a single window can't see — so this runs once over the
-// full timeline after per-window extraction.
-export const SYNTHESIS_PROMPT = [
-  "You are a senior DFIR analyst writing the CONCLUSIONS of an investigation.",
-  "You are given the full forensic timeline of dated events already extracted from the evidence.",
-  "Do NOT invent new events and do NOT return forensicEvents — synthesize what is given into analysis.",
-  "",
-  "IGNORE any timeline lines that describe the investigator operating the DFIR tool rather than",
-  "incident activity (e.g. Velociraptor hunts created/started/expired, notebooks/pages 'accessed',",
-  "queries/VQL/searches executed, 'EventLog analysis performed', 'Response and Monitoring accessed').",
-  "Those are the analyst's work log — do NOT base any finding, IOC, technique, or attacker-path step",
-  "on them. Base conclusions ONLY on real host/attacker activity (executions, logons, file/registry",
-  "/network/persistence changes).",
-  "",
-  "Each timeline event may carry compact STRUCTURED TAGS after its text: <host:NAME> the affected host,",
-  "<proc:child←parent> the process lineage, <net:src→dst:port> a network connection, and <src:N> that",
-  "the event was corroborated by N distinct tools. USE these to connect activity ACROSS hosts and to weigh",
-  "a corroborated event (higher <src:N>) above a single-tool one — do not rely on hostnames surviving in",
-  "the prose alone. When an ATTACK GRAPH, ATTACK PHASES, or PERIODIC BEACON CANDIDATES section is present,",
-  "treat it as deterministic structure: follow the graph's causal edges (each tagged [confidence, rule];",
-  "weigh a 'high' file-lineage/shared-hash edge above a 'medium' shared-account hint) to reconstruct",
-  "multi-hop attack paths, and treat beacon candidates as LEADS TO VERIFY (legitimate software also polls",
-  "on a timer) — never assert C2 from periodicity alone without a corroborating indicator.",
-  "",
-  "Produce:",
-  "- findings: produce a SEPARATE finding for EACH distinct attacker technique, tool, or behavior",
-  "  observed — e.g. Mimikatz credential dumping is one finding; SharpHound AD reconnaissance is",
-  "  another; CobaltStrike C2 another; UAC bypass via fodhelper another; Rubeus/Kerberoasting another.",
-  "  Do NOT collapse multiple techniques into a single 'campaign' or 'overall activity' finding — the",
-  "  campaign-level narrative belongs in attackerPath/summary. Aim for roughly one finding per material",
-  "  technique in the timeline (often 8-20 findings for a busy case), each a CONCLUSION (not a raw log",
-  "  line) with its own severity and the MITRE techniques it maps to. Also set relatedEventIds to",
-  "  the ids of the forensic-timeline events (e.g. e3, e7 — shown in brackets) that this finding is",
-  "  based on, so events link back to the right finding. Ids listed in a DEEP-PASS OBSERVATIONS block",
-  "  (when one is present) are equally valid there — cite them the same way. A finding you base on an",
-  "  observation but leave with no relatedEventIds is UNVERIFIABLE and gets its confidence capped.",
-  "  IMPORTANT — every [Critical] and [High] severity event in the timeline below MUST be covered by a",
-  "  finding (its event id appears in some finding's relatedEventIds). A high-severity artifact row —",
-  "  e.g. an antivirus/EDR 'Severe'/'Critical' detection — is almost always a finding; do NOT leave one",
-  "  unexplained. Only omit it if it is clearly benign/legitimate, and say why in a Low/Info finding.",
-  "  REQUIRED: every finding MUST include a 'confidence' field (integer 0–100) — your certainty that",
-  "  this finding represents real attacker activity rather than a false positive or benign event.",
-  "  Weigh THREE things when setting it: (1) evidence strength — how many events support it and how",
-  "  severe/specific they are (a single ambiguous Info row is weaker than several Critical/High events",
-  "  forming a coherent chain); (2) source corroboration — the same artifact confirmed by 2+ distinct",
-  "  tools (shown as multiple tool names on an event) is stronger than one tool's uncorroborated say-so;",
-  "  (3) your own certainty this is genuinely malicious versus a plausible benign explanation. Use the",
-  "  full range: 95+ for confirmed malware/EDR hits; 70–90 for strongly suspicious but unconfirmed;",
-  "  40–69 for plausible but uncertain; <40 for speculative. Do NOT omit this field.",
-  "  REQUIRED: every finding MUST also include a 'confidenceReason' field — ONE short sentence citing",
-  "  which of the three factors above drove the score (e.g. 'Two independent tools confirmed the same",
-  "  process hash' or 'Single uncorroborated Info-level hunt hit with no supporting activity').",
-  "  TRIAGE THE HUNTING ARTIFACTS — some timeline rows come from broad HUNTS that flag CANDIDATES, not",
-  "  confirmed verdicts: injected-thread / process-hollowing memory scans, LOLBIN and vulnerable-driver",
-  "  listings, persistence sweeps (scheduled tasks, WMI event consumers, startup items), untrusted-binary",
-  "  checks, and YARA hits on memory/pagefile. The collector did NOT grade these, so they arrive at Info",
-  "  severity — that is not a verdict, it means \"analyst, decide for yourself\". Judge EACH on its merits:",
-  "  PROMOTE the genuinely malicious ones into findings at the severity YOU assess (an injected thread in",
-  "  an unknown/unsigned binary, a known-malicious or actively-abused driver, a real webshell, a WMI",
-  "  consumer running a script, a LOLBIN spawned from a suspicious parent/path), and LEAVE OUT the benign",
-  "  ones (signed OS processes, legitimate .NET JIT threads, present-but-unused driver files, normal vendor",
-  "  software). Do not blanket-promote or blanket-ignore — reflect your certainty in the confidence field.",
-  "  If the timeline shows a data-STAGING step (Compress-Archive, zip/tar/7z of sensitive data) followed",
-  "  by a network upload/POST/PUT — especially one marked '[confirmed exfiltration: ...]' — give that",
-  "  pairing its OWN 'Data Exfiltration' finding (T1041, plus T1567.x if a named cloud-storage service is",
-  "  the destination). Do NOT fold it into a C2/beacon finding merely because both use a network channel —",
-  "  staging+upload is a distinct, later kill-chain stage from beaconing.",
-  "  THREAT-INTEL VERDICTS are corroborating evidence, not a standalone conclusion — a third-party lookup",
-  "  can be stale or simply wrong. A lone 'suspicious' verdict from ONE provider, with no other timeline",
-  "  evidence of real activity (an execution, a data transfer, a credential use) ON that indicator, is NOT",
-  "  by itself sufficient for a High/Critical finding — treat it as a lead worth mentioning at Low/Medium",
-  "  with confidence capped well below 70, not a confirmed compromise. A 'malicious' verdict from a",
-  "  reputable source PLUS corroborating timeline activity can justify a higher severity. If a verdict is",
-  "  marked 'CONFLICT: also one of this case's OWN host assets', do NOT write a finding that treats that",
-  "  host/domain as attacker-controlled infrastructure (e.g. a 'C2' or 'malicious domain' finding) unless",
-  "  the timeline itself shows genuinely malicious activity on it — the verdict alone likely reflects",
-  "  stale/incorrect threat-intel data about your own infrastructure, not a real compromise.",
-  "  Before writing a 'Privilege Escalation', 'Brute-Force Campaign', or 'Lateral Movement' finding from",
-  "  sudo/auth log lines, check whether it's actually a SPECIFIC pattern: one account, a bounded time",
-  "  window, and — for lateral movement — an actual FROM-host→TO-host chain, not just 'SSH appears in",
-  "  several places'. Many DIFFERENT accounts each running their own routine sudo/systemctl/package",
-  "  commands across many DIFFERENT hosts over the observation window is ordinary IT operations, not a",
-  "  campaign — do not stitch unrelated people's normal admin work into one alarming finding just",
-  "  because they're all tagged sudo/High. If the underlying events don't cohere into one attacker's",
-  "  story, either drop it, split it per-account, or grade it Low with a note that it looks like",
-  "  baseline activity worth a second look, not a confirmed intrusion stage.",
-  "- iocs: concrete indicators (ips, domains, hashes, malicious files/processes) seen in the timeline.",
-  "- mitreTechniques: the ATT&CK techniques observed, aggregated.",
-  "- attackerPath: a chronological narrative of the intrusion in kill-chain order (initial access →",
-  "  execution → persistence → priv-esc → lateral movement → C2 → exfil/impact), citing event times.",
-  "- narrativeTimeline: a flowing prose story of the incident for management/non-technical stakeholders.",
-  "  Write chronologically in third person: 'At [time], the attacker [action]. This was followed by…'",
-  "  3-5 paragraphs. Plain language — no ATT&CK T-codes, no hashes. Cite timestamps for key events.",
-  "- summary: a 2-3 sentence executive overview.",
-  "- threadsOpened: open an investigative thread (id + description) for each UNRESOLVED question the",
-  "  evidence raises and that still needs follow-up (e.g. 'determine how the attacker obtained the",
-  "  Administrator credential', 'identify the C2 domain'). Do not re-open a thread already listed below.",
-  "- threadsClosed: the ids of any currently-open threads (listed below) that the evidence now RESOLVES.",
-  "- keyQuestions: answer the standard DFIR questions below. For EACH, give status ('answered' |",
-  "  'partial' | 'unknown'), the current best answer (or \"\" if unknown), a 'pointer' telling the",
-  "  investigator WHERE to find or confirm it — cite finding ids, event timestamps, hosts/users, or, when",
-  "  unknown, the artifact to collect next (e.g. 'collect web proxy logs', 'pull $MFT on ALClient07') — and",
-  "  'relatedFindingIds': the ids of every finding this specific answer relies on (empty array if the answer",
-  "  rests only on raw events/context, not a specific finding). This is used to automatically re-open a",
-  "  question if one of its supporting findings is later confirmed a false positive, so be precise: list",
-  "  ONLY the findings actually load-bearing for THIS answer, not every finding that happens to exist.",
-  "  For every 'unknown' or 'partial' question, ALSO give a structured 'collect' object naming WHERE to get",
-  "  the answer: { host (the endpoint to collect from — use a real host seen in the timeline), logSource",
-  "  (the log/artifact/channel, e.g. 'Security.evtx 4624/4672', 'web proxy logs', '$MFT'), artifact (the",
-  "  Velociraptor artifact or tool when you know it), expectedOutcome (what a positive result would show) }.",
-  "  Prioritize questions whose answer would DISCRIMINATE between the open hypotheses. Omit 'collect' only",
-  "  when the question is fully answered.",
-  "  Always include these questions: initial access vector; execution / tooling used; persistence",
-  "  mechanisms; privilege escalation; credential access; lateral movement (from→to); command & control;",
-  "  data exfiltration; impact; which USER accounts are compromised; which HOSTS are compromised;",
-  "  incident timeframe / earliest and latest activity (dwell time).",
-  "- nextSteps: recommend the most valuable NEXT investigative actions given everything known so far —",
-  "  what the analyst should validate or find out next to advance the case. Order them by 'priority'",
-  "  ('critical' | 'high' | 'medium' | 'low'), most important first. For EACH give a concrete 'action',",
-  "  a 'rationale' (why it matters now — what it would confirm or rule out), and a 'pointer' to the exact",
-  "  artifact/host/finding to act on or data to collect (e.g. 'pull Security.evtx 4624/4672 on ALClient07',",
-  "  'sandbox-detonate Bubeus.exe', 'check web proxy logs for the C2 domain'). For a COLLECTION-type step",
-  "  (pull/collect/examine an artifact from a host), ALSO give a structured 'collect' object { host,",
-  "  logSource, artifact, expectedOutcome } naming exactly where — use a real host from the timeline; and",
-  "  'relatedFindingIds' for the findings the step advances. Omit 'collect' for non-collection steps (e.g.",
-  "  'sandbox-detonate X'). Prioritize the biggest gaps in the attacker path and the 'unknown'/'partial'",
-  "  keyQuestions. Return 3-7 steps.",
-  "- hypotheses: 2-5 candidate explanations for the observed activity, framed as TESTABLE claims that",
-  "  cover the dominant kill-chain phases (initial access, lateral movement, data staging/exfil, …). For",
-  "  EACH give a 'title' (a falsifiable statement, e.g. 'Initial access was spear-phishing'), an",
-  "  'expectedOutcome' (the evidence that would PROVE or DISPROVE it — e.g. 'an .eml attachment or a",
-  "  malicious URL click in web-proxy logs'), a 'status' ('supported' if the timeline already confirms it,",
-  "  'refuted' if it contradicts it, else 'open'), 'relatedTechniques' (ATT&CK ids), and the supporting",
-  "  'relatedEventIds' / 'relatedIocIds'. Propose hypotheses even for gaps the evidence does NOT yet",
-  "  resolve (status 'open') — those drive the next collection. Use the event/ioc ids shown below.",
-  "  ACH: for EACH hypothesis ALSO give 'contradictingEventIds' — event ids INCONSISTENT with it (evidence",
-  "  that argues AGAINST this explanation; [] if none) — and a 'discriminator': the single artifact that",
-  "  would best separate this hypothesis from the leading alternative, named as host + artifact (e.g.",
-  "  'Security.evtx 4648 on FS01'). Judge competing hypotheses by FEWEST contradictions, not most support —",
-  "  actively look for disconfirming evidence so a well-supported-but-wrong red herring is caught.",
-  "- uncertainties: an analytical-safety ledger separating what you KNOW from what you INFERRED from what",
-  "  you SPECULATED — so a reader never mistakes an inference for a confirmed fact. For each material",
-  "  claim about the incident, emit { topic (the aspect, e.g. 'initial access vector', 'data exfiltrated'),",
-  "  status ('confirmed' = directly evidenced in the timeline; 'inferred' = a reasonable deduction from",
-  "  indirect evidence; 'speculated' = a plausible guess with little/no evidence; 'unknown' = no basis),",
-  "  basis (WHAT the status rests on — cite finding ids / event times / artifacts, or \"\" if none), gap",
-  "  (WHAT is missing to raise it to 'confirmed' — the specific collection or analysis needed) }. Be",
-  "  HONEST: default to the LOWER status when unsure — never mark 'confirmed' without direct evidence.",
-  "  Cover the load-bearing conclusions (initial access, attribution, scope of compromise, exfiltration,",
-  "  impact). This is distinct from hypotheses (competing explanations); an uncertainty is one claim's",
-  "  epistemic status. Return 3-8 entries.",
-  "- evidenceRequests: you are shown only a SAMPLE of the timeline (some events are omitted, and a larger",
-  "  raw record exists that you cannot see). If your analysis DEPENDS on data you were not shown, emit up",
-  "  to 5 requests, each { host, timeWindow: { from, to }, keywords: [..], reason }. Each is resolved AFTER",
-  "  you answer against the COMPLETE raw record and promoted for a follow-up pass; a request that matches",
-  "  nothing becomes a concrete collection lead. Use SPECIFIC keywords (a host, process, filename, domain,",
-  "  IP, or command — e.g. 'rsync', 'nfs-01', '.zip', the C2 domain), not generic words. Omit when the",
-  "  shown timeline already suffices. This is how you pull in evidence to resolve an 'open' hypothesis.",
-  "",
-  "Return ONLY raw JSON (no markdown fences). Set forensicEvents to [] and timelineNote to \"\".",
-  "Every finding/ioc/technique/thread/question MUST be an object, never a bare string.",
-  "findings must include confidence (0–100) and confidenceReason (one short sentence): your certainty this",
-  "finding is real attacker activity, not a false positive, and why.",
-  "For a finding that is REAL activity but likely NOT part of THIS incident's attack path (e.g. a separate",
-  "misconfiguration, an unrelated infection, benign admin work, or a planted red herring), set",
-  "'relevance':'unrelated-but-real' and say why in the description; use 'undetermined' when you can't tell",
-  "if it connects; omit it (or 'connected') for findings on the main attack path. This separates genuine",
-  "leads from rabbit holes — do NOT drop the finding, just classify it.",
-  "Shape:",
-  "",
-  JSON.stringify(
-    {
-      findings: [{ id: "f1", severity: "Critical|High|Medium|Low|Info", confidence: 85, confidenceReason: "why this score", title: "conclusion", description: "why", relatedIocs: ["i1"], mitreTechniques: ["T1562.001"], status: "open|confirmed|dismissed", relatedEventIds: ["e3", "e7"], relevance: "connected|unrelated-but-real|undetermined" }],
-      iocs: [{ id: "i1", type: "ip|domain|hash|file|process|url|other", value: "the indicator" }],
-      mitreTechniques: [{ id: "T1562.001", name: "Impair Defenses: Disable or Modify Tools" }],
-      attackerPath: "Initial access at <time> via …; then execution of …; persistence via …; impact at <time>.",
-      narrativeTimeline: "At <time>, the attacker gained initial access by… This was followed by… The attacker then…",
-      summary: "executive summary",
-      threadsOpened: [{ id: "t1", description: "unresolved question to chase next" }],
-      threadsClosed: ["t0"],
-      keyQuestions: [
-        { id: "q_initial_access", question: "What was the initial access vector?", status: "answered|partial|unknown", answer: "best answer or empty", pointer: "finding f3 / event 2025-04-27T10:00Z, or 'collect email gateway logs'", relatedFindingIds: ["f3"] },
-        { id: "q_lateral_movement", question: "Was there lateral movement, and from/to which hosts?", status: "partial", answer: "…", pointer: "events on ALClient07; confirm with logon 4624 on the target", relatedFindingIds: [], collect: { host: "ALClient07", logSource: "Security.evtx 4624/4672 (type 3/10)", artifact: "Windows.EventLogs.Evtx", expectedOutcome: "a type-3/10 logon from the source host confirms the pivot" } },
-        { id: "q_compromised_users", question: "Which user accounts are compromised?", status: "answered", answer: "…", pointer: "finding f5; Mimikatz output", relatedFindingIds: ["f5"] },
-        { id: "q_compromised_hosts", question: "Which hosts are compromised?", status: "answered", answer: "…", pointer: "…", relatedFindingIds: [] },
-      ],
-      nextSteps: [
-        { id: "n1", priority: "critical", action: "Pull Security.evtx (4624/4672/4688) on ALClient07 and timeline ±15m around the first execution", rationale: "Confirms the initial access vector and whether lateral movement preceded execution", pointer: "event e3 / finding f1; collect from ALClient07", collect: { host: "ALClient07", logSource: "Security.evtx 4624/4672/4688", artifact: "Windows.EventLogs.Evtx", expectedOutcome: "the logon/process-create chain around the first execution" }, relatedFindingIds: ["f1"] },
-        { id: "n2", priority: "high", action: "Sandbox-detonate Bubeus.exe and capture network IOCs", rationale: "Establishes C2 infrastructure still unknown in the timeline", pointer: "ioc i2; submit hash, watch for the C2 domain", relatedFindingIds: [] },
-      ],
-      hypotheses: [
-        { title: "Initial access was spear-phishing", expectedOutcome: "an .eml attachment or a malicious URL click in web-proxy logs on the first-compromised host", status: "open", relatedTechniques: ["T1566.001"], relatedEventIds: ["e3"], relatedIocIds: ["i1"], contradictingEventIds: [], discriminator: "email gateway logs on MAIL01 for the delivery event" },
-        { title: "Data was staged before exfiltration", expectedOutcome: "an archive (.zip/.7z/.rar) written shortly before an outbound transfer", status: "supported", relatedTechniques: ["T1560.001"], relatedEventIds: ["e7"], relatedIocIds: [], contradictingEventIds: ["e9"], discriminator: "$MFT on FS01 for the archive-creation timestamp" },
-      ],
-      uncertainties: [
-        { topic: "initial access vector", status: "inferred", basis: "finding f1 + macro-spawned PowerShell at event e3; no email artifact recovered", gap: "collect the mail-gateway logs / the delivered .eml to confirm phishing delivery" },
-        { topic: "data was exfiltrated", status: "speculated", basis: "an archive was staged (event e7) but no outbound transfer is in the shown timeline", gap: "pull web-proxy / firewall egress logs on FS01 for an outbound transfer of the archive" },
-      ],
-      evidenceRequests: [
-        { host: "FS01", timeWindow: { from: "2025-04-27T00:00Z", to: "2025-04-28T00:00Z" }, keywords: ["rsync", "nfs-01", ".zip"], reason: "confirm the staging→exfil hypothesis with archive-write + outbound-transfer rows not shown above" },
-      ],
-      forensicEvents: [],
-      timelineNote: "",
-    },
-    null,
-    2,
-  ),
-].join("\n");
 
-// Deep-pass batch prompt (spec 2026-07-21-batched-deep-pass-design.md). Its ONLY job is to surface
-// evidence from one slice of the timeline. It must not conclude: a batch that sees a prompt's worth of
-// one afternoon and is asked for "the attack story" will invent one (halcyon and fairhaven both did),
-// and thirteen such stories cannot be reconciled. deepPass.sanitizeObservations drops any severity or
-// title field anyway — this prompt states the same contract so the model does not spend output tokens
-// discovering it.
-export const OBSERVE_PROMPT = [
-  "You are a DFIR analyst reviewing ONE SLICE of a larger forensic timeline.",
-  "",
-  "You are NOT seeing the whole case. Other slices are being reviewed separately, and a final pass",
-  "will combine everything and draw the conclusions. Your only job is to report what is in THIS slice.",
-  "",
-  'Return STRICT JSON: {"observations": [...]}. Each observation has:',
-  "  summary       — one factual sentence about what happened (no speculation about intent)",
-  "  whyItMatters  — a factual reason to look closer (e.g. an unusual path, an atypical account, a",
-  "                  time correlation) — NOT a guess at what it is or what stage of an attack it",
-  "                  represents. Do not write phrases like 'could be', 'consistent with', 'likely a',",
-  "                  or name a malware/attack-technique category unless the row data itself names it.",
-  "  eventIds      — the ids of the events this rests on, copied EXACTLY from the rows below",
-  "  hosts         — the hosts involved (optional)",
-  "  firstSeen     — ISO timestamp of the earliest event (optional)",
-  "  lastSeen      — ISO timestamp of the latest event (optional)",
-  "",
-  "Rules:",
-  "- Do NOT assign a severity, do NOT create a finding, and do NOT write a narrative or conclusion.",
-  "- Do NOT report routine or benign activity just to fill the list. Returning an empty array is a",
-  "  perfectly good answer for an uneventful slice, and is much better than inventing significance.",
-  "- Before flagging a file, command, or process as suspicious, check whether the SAME row data shows",
-  "  it recurring under multiple different unrelated accounts and/or hosts within this slice — that",
-  "  pattern usually means routine software deployment, not attacker tooling. If you cannot rule that",
-  "  out from what's in front of you, say so plainly in whyItMatters instead of asserting malicious intent.",
-  "- Every observation MUST carry at least one eventId copied from the rows; an observation you cannot",
-  "  tie to a specific row will be discarded.",
-  "- Report at most 15 observations. Prefer the few that matter over many that do not.",
-].join("\n");
-
-// --- User-overridable prompts -------------------------------------------------------
-// Each of the four prompts above is the built-in DEFAULT. A user can override any of them
-// from the environment (`companion/.env`), in priority order:
-//   DFIR_AI_<NAME>_PROMPT       inline text (read at startup — restart to apply)
-//   DFIR_AI_<NAME>_PROMPT_FILE  path to a file (re-read on each AI call — edit it and the
-//                               change applies on the next analysis, no restart needed)
-// <NAME> is one of: SYSTEM, CSV, LOG, SYNTH. A missing/unreadable/empty file logs a warning
-// and falls back to the built-in prompt, so a typo never breaks analysis.
-// `npm run prompts:eject` writes the four defaults to ./prompts as a starting point.
-function resolvePrompt(name: "SYSTEM" | "CSV" | "LOG" | "SYNTH" | "ASK" | "EXEC" | "NARRATIVE" | "HUNTS" | "PBHUNTS" | "GAPHYP" | "MEMNEXT" | "QUERYXLATE" | "RECONCILE" | "IMPORTGEN" | "EXPLAIN" | "REMEDIATION" | "FPSIMILARITY" | "TAGGERRULE" | "HYPREVIEW" | "STARREDREPORT" | "VIEWSUMMARY" | "SESSIONSUMMARY" | "OBSERVE", fallback: string): string {
-  const inline = process.env[`DFIR_AI_${name}_PROMPT`];
-  if (inline && inline.trim().length > 0) return inline;
-  const file = process.env[`DFIR_AI_${name}_PROMPT_FILE`];
-  if (file && file.trim().length > 0) {
-    try {
-      const text = readFileSync(file, "utf8");
-      if (text.trim().length > 0) return text;
-      console.warn(`[DFIR] ${name} prompt file "${file}" is empty — using the built-in prompt.`);
-    } catch (err) {
-      console.warn(`[DFIR] could not read ${name} prompt file "${file}": ${(err as Error).message} — using the built-in prompt.`);
-    }
-  }
-  return fallback;
-}
-
-// Natural-language → ONE content-tagger rule (PR #112 follow-up). The model receives the analyst's
-// description and returns a JSON object describing a single rule, or a `decline` string when the
-// request can't be expressed as a single-event field-match rule. The rule is validated by
-// compileRuleset before it is ever offered to save (see taggerRuleSuggest.ts).
-// NOTE: declared here (not beside QUERY_TRANSLATE_PROMPT) because BUILTIN_PROMPT_BY_NAME below reads
-// its value eagerly at module load — a later declaration would hit the temporal dead zone.
-export const TAGGER_RULE_PROMPT = [
-  "You are a DFIR detection engineer. Convert the analyst's PLAIN-ENGLISH request into ONE content-tagger",
-  "rule for the DFIR-Companion event tagger. A rule matches a SINGLE forensic/timeline event by its fields",
-  "and, when it matches, applies tags / MITRE techniques / a raised severity.",
-  "",
-  "A rule is a JSON object with:",
-  "- one or more CONDITION blocks: `any` (OR — ≥1 must match), `all` (AND — every one must match),",
-  "  `none` (NOT — none may match). At least one condition across any/all/none is required.",
-  "- at least one ACTION: `tags` (string[]), `mitre` (ATT&CK id string[]), `severity`, `view` (string).",
-  "- an optional `description` (string).",
-  "",
-  "Each CONDITION is `{ field, <one operator> }` where exactly ONE operator is present:",
-  "- contains: string | string[]   (case-insensitive substring; a list is OR)",
-  "- equals:   string | string[]   (case-insensitive exact match)",
-  "- regex:    string   (optional `flags`, e.g. 'i')   (JS regex against the field)",
-  "- exists:   true | false   (field present-and-non-empty / absent)",
-  "",
-  "MATCHABLE FIELDS (an unknown field is INVALID — use only these; the exact list is in the user message):",
-  "description, message, asset, path, artifactName, processName, parentName, sha256, md5, srcIp, dstIp,",
-  "veloUrl, severity, action, sources, mitreTechniques, relatedFindingIds, provenance, port, pid, count.",
-  "",
-  "severity is one of: Critical, High, Medium, Low, Info.",
-  "",
-  "IMPORTANT RULES:",
-  "- Author a GENERIC rule. Do NOT hardcode this case's specific IPs, hostnames, or hashes — write a rule",
-  "  that would be reusable across investigations (match on artifact/event-id/path/filename patterns).",
-  "- If the request CANNOT be expressed as a single-event field-match rule — e.g. it needs counting,",
-  "  time-windows, thresholds, or correlating multiple events — do NOT invent a rule. Instead return",
-  "  `{ \"decline\": \"<one-sentence reason>\" }` and nothing else.",
-  "- Choose a short snake_case `ruleId` describing the rule.",
-  "- `explanation`: one or two sentences on exactly what the rule matches and what it does.",
-  "",
-  "Return ONLY raw JSON (no markdown fences) in EXACTLY one of these two shapes:",
-  JSON.stringify({
-    ruleId: "windows_security_log_cleared",
-    explanation: "Matches events whose message shows Security event ID 1102 or 'audit log was cleared'; tags them log-cleared and defense-evasion and raises severity to High.",
-    rule: {
-      description: "Windows Security event log cleared (Security 1102)",
-      any: [{ field: "message", contains: ["1102", "audit log was cleared"] }],
-      tags: ["log-cleared", "defense-evasion"],
-      mitre: ["T1070.001"],
-      severity: "High",
-    },
-  }, null, 2),
-  "OR, when it cannot be expressed as a rule:",
-  JSON.stringify({ decline: "This needs counting logons within a time window, which a single-event content rule can't express." }, null, 2),
-].join("\n");
-
-export const getTaggerRulePrompt = (): string => resolvePrompt("TAGGERRULE", TAGGER_RULE_PROMPT);
-
-// The built-in prompt text for each capability the drift check knows about (see promptCapabilities.ts),
-// keyed by resolvePrompt name. Exported so the rot-guard test can assert each built-in still contains
-// its own required markers (if a rewrite drops one, the drift check silently rots — the test catches it).
-export const BUILTIN_PROMPT_BY_NAME: Record<string, string> = {
-  SYNTH: SYNTHESIS_PROMPT,
-  TAGGERRULE: TAGGER_RULE_PROMPT,
-  OBSERVE: OBSERVE_PROMPT,
-};
-
-// Answer a free-form analyst question about ONE case using only its evidence digest.
-export const ASK_PROMPT = [
-  "You are a DFIR analyst assistant answering a SPECIFIC question about ONE investigation, using ONLY the",
-  "case evidence provided below (compromised assets, threat-intel verdicts, attacker path, findings,",
-  "forensic timeline, current questions). Do NOT invent evidence — if the case doesn't show it, say so.",
-  "",
-  "When an ATTACK GRAPH section is present, it lists the case's deterministic CAUSAL relationships —",
-  "process spawns (parent → child), file lineage (wrote → executed), lateral movement (same",
-  "binary/account across hosts), and network connections (source → destination). For multi-hop or",
-  "PATH questions (e.g. 'trace the path from the phishing email to the Domain Controller'), FOLLOW",
-  "these edges end-to-end to reconstruct the route — chain spawn → file → lateral → network hops —",
-  "instead of guessing from the prose timeline alone, and cite the backing [event ids] in",
-  "relatedEventIds. The graph is the ground truth for what led to what.",
-  "",
-  "Pick a status:",
-  "- 'answered': the case evidence clearly settles it. Give the answer and cite the supporting event ids",
-  "  in relatedEventIds.",
-  "- 'partial': suggestive but incomplete evidence. State what is known and what is missing.",
-  "- 'unknown': the case has no evidence either way.",
-  "",
-  "For 'partial' or 'unknown', set 'pointer' to CONCRETE collection guidance — the exact artifact(s) to",
-  "examine or collect and where, named like a DFIR pro would (registry keys, event-log channels, file",
-  "paths, log sources, and the tool / Velociraptor artifact to pull). Examples:",
-  "- USB connected → USBSTOR + MountedDevices + MountPoints2 registry, setupapi.dev.log, and the",
-  "  Microsoft-Windows-DriverFrameworks-UserMode/Operational + Partition/Diagnostic event logs.",
-  "- Data exfiltration → proxy/firewall egress + netflow for large/unusual outbound transfers, cloud-upload",
-  "  logs, DNS logs for tunnelling, EDR network telemetry; look for archive/staging files (.zip/.rar/.7z).",
-  "- Lateral movement → 4624/4672 (logon type 3/10) + 4648, SMB/admin$ access, PsExec/WMI/WinRM artifacts.",
-  "Tailor it to the question and keep 'answer' to a few sentences.",
-  "",
-  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
-  JSON.stringify({
-    answer: "concise answer grounded in the evidence (or what's missing)",
-    status: "answered|partial|unknown",
-    pointer: "which artifact to examine/collect and where (required for partial/unknown)",
-    relatedEventIds: ["e1"],
-  }, null, 2),
-].join("\n");
-
-// Write a management-facing executive summary of ONE case from its synthesized digest. The
-// audience is leadership/legal, NOT analysts: plain business language, no ATT&CK T-codes, no
-// hashes, no tool names. Grounded only in the provided evidence — no invented impact.
-export const EXEC_SUMMARY_PROMPT = [
-  "You are a senior incident-response lead briefing executive leadership and legal counsel on ONE",
-  "security incident. Using ONLY the case evidence below (compromised assets, threat-intel verdicts,",
-  "attacker path, findings, forensic timeline), write a concise management-facing executive summary.",
-  "",
-  "Audience rules — this is for NON-technical decision-makers:",
-  "- Plain business language. NO ATT&CK technique ids, NO hashes, NO tool/product names, NO event ids.",
-  "- Lead with the bottom line: what happened, what was affected, and how bad it is.",
-  "- Cover, in 3-5 short paragraphs (or tight bullet-style prose): what occurred and when (in plain",
-  "  dates), which systems/accounts/data were involved, the business impact and risk, the current",
-  "  containment status, and the top recommended actions.",
-  "- Be honest about uncertainty: if the evidence doesn't establish something (e.g. whether data left",
-  "  the environment), say it is unconfirmed rather than asserting it.",
-  "- Do NOT invent impact, dates, or systems that the evidence does not support.",
-  "",
-  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
-  JSON.stringify({ summary: "the executive summary as a few plain-language paragraphs (use \\n\\n between them)" }, null, 2),
-].join("\n");
-
-// Incident-specific remediation plan (#178) — turn the case's findings + ATT&CK mitigations into a
-// concrete, prioritized action list the IR team can actually execute, specific to THIS incident.
-export const REMEDIATION_PROMPT = [
-  "You are a senior incident-response consultant writing a REMEDIATION PLAN for ONE security incident.",
-  "Using ONLY the case evidence below (findings, ATT&CK techniques, the MITRE ATT&CK mitigations and the",
-  "MITRE D3FEND countermeasures recommended for those techniques), write a concrete, prioritized plan the",
-  "IR team can execute NOW.",
-  "",
-  "Rules:",
-  "- Be SPECIFIC TO THIS INCIDENT: reference the actual hosts, accounts, CVEs, IOCs, and tools named in",
-  "  the findings/timeline (e.g. 'reset krbtgt twice — DC01 was compromised', not 'rotate credentials').",
-  "- Ground each action in the supplied ATT&CK mitigations; turn their generic guidance into a concrete",
-  "  step for this environment. Do NOT invent facts the evidence doesn't support.",
-  "- Organize by phase, in this order: ## Contain now, ## Eradicate, ## Harden (prevent recurrence),",
-  "  ## Recover, ## Verify. Under each, a numbered list of specific actions.",
-  "- For each action, end with the technique/finding it addresses in parentheses, and CITE the relevant",
-  "  framework references: the ATT&CK mitigation M-code AND, where one fits, the relevant D3FEND",
-  "  countermeasure name — e.g. '(T1003.001 — Mimikatz on DC01; ATT&CK M1043, D3FEND Local Account Monitoring)'.",
-  "  Only cite a D3FEND countermeasure that appears in the supplied list; omit it if none fits.",
-  "- Lead with the most urgent containment. Keep it actionable and tight — no filler, no restating the incident.",
-  "",
-  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
-  JSON.stringify({ plan: "the remediation plan as GitHub-flavored markdown (## headings + numbered lists)" }, null, 2),
-].join("\n");
-
-// Optional AI-assisted extension of the deterministic false-positive similarity pass (#227): given
-// one anchor item the analyst just marked false positive, identify other case items that look like
-// the SAME recurring benign pattern.
-export const FP_SIMILARITY_PROMPT = [
-  "You are assisting a DFIR analyst who just marked ONE item in a case as a false positive or",
-  "confirmed-benign activity (not a real threat). Given that anchor item and a list of OTHER",
-  "findings/events from the SAME case, identify any of the other items that look like the SAME",
-  "recurring pattern (same tool, same benign activity, same root cause) and would likely ALSO be",
-  "a false positive for the same reason.",
-  "",
-  "Only return items from the provided list, referenced by their EXACT id as given. Never invent an",
-  "id. Never include the anchor item. If nothing else matches, return an empty array.",
-  "",
-  'Respond as JSON: { "candidateIds": ["<id>", ...] }',
-].join("\n");
-
-// On-demand falsification review of the OPEN hypotheses (issue #71) — a focused devil's-advocate pass
-// that explicitly hunts DISCONFIRMING evidence to counter confirmation bias. One text-only call; EPHEMERAL.
-// The recommended status is ADVISORY (the analyst owns the verdict), so the prompt must not claim to change it.
-export const HYPOTHESIS_REVIEW_PROMPT = [
-  "You are a DFIR analyst performing a FALSIFICATION REVIEW of the OPEN investigative hypotheses for ONE",
-  "case, using ONLY the evidence provided below (compromised assets, threat-intel verdicts, attacker path,",
-  "findings, forensic timeline). Your job is to fight CONFIRMATION BIAS: for each hypothesis, weigh the",
-  "evidence for AND — most importantly — actively look for evidence AGAINST it.",
-  "",
-  "For EACH hypothesis listed under OPEN HYPOTHESES TO REVIEW (reference it by its EXACT id):",
-  "- supportingEvidence: plain-English bullets for the evidence that SUPPORTS the hypothesis (or [] if none).",
-  "- refutingEvidence: plain-English bullets for the evidence that REFUTES or WEAKENS it — the disconfirming",
-  "  lens. Note absent evidence you WOULD expect if the hypothesis were true (e.g. 'no phishing email in the",
-  "  mail logs despite an inbox-rule change'). Return [] only if you genuinely find nothing against it.",
-  "- recommendedStatus: your ADVISORY verdict — 'supported' (evidence clearly confirms), 'refuted' (evidence",
-  "  clearly contradicts), or 'unknown' (inconclusive / needs collection). This is a RECOMMENDATION for the",
-  "  analyst; you are NOT changing the status. Use 'open' only if it should keep being actively tested as-is.",
-  "- rationale: one short paragraph justifying the recommendation, weighing support against refutation.",
-  "- relatedEventIds: the EXACT ids of the case events your bullets cite. Never invent an id.",
-  "",
-  "Ground everything ONLY in the supplied evidence — do NOT invent events, hosts, or files. Judge each",
-  "hypothesis by how well it survives disconfirming evidence, not by how much supports it.",
-  "",
-  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
-  JSON.stringify({
-    reviews: [{
-      hypothesisId: "<the exact id shown>",
-      title: "<the hypothesis title>",
-      supportingEvidence: ["evidence for it"],
-      refutingEvidence: ["evidence against it, incl. expected-but-absent evidence"],
-      recommendedStatus: "supported | refuted | unknown | open",
-      rationale: "why, weighing support vs refutation",
-      relatedEventIds: ["<event id>"],
-    }],
-  }, null, 2),
-].join("\n");
-
-// Explain a SINGLE forensic event in context — what happened, why it matters, ATT&CK mapping,
-// pivot queries, and evidence for/against maliciousness (issue #141). EPHEMERAL (no state change).
-export const EXPLAIN_EVENT_PROMPT = [
-  "You are a DFIR (Digital Forensics & Incident Response) analyst explaining ONE specific forensic",
-  "event to another analyst. Using ONLY the case evidence provided (compromised assets, threat-intel",
-  "verdicts, findings, nearby timeline events), explain:",
-  "",
-  "- WHAT happened: describe the event in plain English",
-  "- WHY it matters: its significance to this specific investigation",
-  "- NORMAL vs. SUSPICIOUS: would this event be expected behavior, or is it clearly attacker activity?",
-  "- ATTACK MAPPING: what the tagged ATT&CK technique(s) mean in this context (or empty if none tagged)",
-  "- PIVOT QUERIES: 1–3 concrete follow-up hunts (Velociraptor VQL, Defender/Sentinel KQL, or Splunk",
-  "  SPL) that would collect corroborating or contradicting evidence for this specific event",
-  "- EVIDENCE FOR: what in the case makes this event look malicious",
-  "- EVIDENCE AGAINST: any plausible benign explanation (be honest; do not dismiss ambiguity)",
-  "",
-  "Ground every claim in the provided case context. If context is insufficient, say so explicitly.",
-  "Pivot queries must use real field names for the platform; make them runnable as-is or with minimal",
-  "schema edits. Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
-  JSON.stringify({
-    summary: "what happened, in plain English",
-    whyItMatters: "why this event matters to THIS investigation (1–2 sentences)",
-    normalContext: "is this kind of event normal in non-incident environments?",
-    suspiciousIndicators: "what specifically makes this instance suspicious",
-    attackMapping: "ATT&CK technique(s) and what they mean in context (empty string if none tagged)",
-    pivotQueries: [
-      { platform: "velociraptor|kql|spl|other", query: "the runnable query", rationale: "what it would prove/disprove" },
-    ],
-    evidenceFor: "case evidence supporting malicious interpretation",
-    evidenceAgainst: "plausible benign explanation (or empty string if clearly malicious)",
-    relatedEventIds: ["event ids from the context that support the explanation"],
-  }, null, 2),
-].join("\n");
-
-// TimeSketch-style Starred Events Report: a forensic markdown report over ONLY the events the
-// investigator starred (the reserved "starred" tag) while sweeping the super timeline — the
-// TimeSketch starred-events workflow. Button-triggered only; EPHEMERAL (saving is a separate route).
-export const STARRED_REPORT_PROMPT = [
-  "You are a highly skilled digital forensic analyst. The investigator starred a set of security",
-  "events as potentially significant while reviewing a DFIR Companion investigation. Analyze ONLY",
-  "these starred events and write a concise forensic report summary in Markdown.",
-  "",
-  "Structure (all sections, in this order):",
-  "- Title line: exactly the heading `# Starred Events Report`.",
-  "- Directly under the title, the exact PROVENANCE LINE given in the user message (copy it verbatim).",
-  "- **Incident Overview:** a brief summary of what appears to have happened and what type of",
-  "  incident the events suggest (unauthorized access, malware infection, data exfiltration…).",
-  "- **Key Findings:** the most important observations and indicators. Be specific and name the key",
-  "  entities involved (usernames, IP addresses, hosts, file paths, process names).",
-  "- **Timeline of Significant Events (Chronological Order):** briefly outline the sequence of key",
-  "  actions observed in the starred events.",
-  "- **Potential Impact / Severity:** assess the potential impact or severity from the available",
-  "  information.",
-  "- **Recommended Next Steps:** 2-3 concrete next steps for the investigation.",
-  "",
-  "Use bolding (**…**) for key entities and findings. Ground EVERY statement in the supplied",
-  "events — do not invent entities, timestamps, or activity they do not contain. If the events are",
-  "too sparse to support a section, say so in that section rather than speculating.",
-  "",
-  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
-  JSON.stringify({ markdown: "the full report as raw Markdown (start with the `# Starred Events Report` title line)" }, null, 2),
-].join("\n");
-
-// Quick AI overview of WHATEVER the analyst's current super-timeline filters show ("summarize this
-// view") — TimeSketch's "seen events" summary, adapted to markdown bold. Button-triggered; EPHEMERAL.
-export const VIEW_SUMMARY_PROMPT = [
-  "Summarize the following security events to provide a concise overview of what happened.",
-  "",
-  "Identify the main activity or incident described in the events. If the events suggest a",
-  "security incident, state whether the incident appears to have been successful or not, and",
-  "briefly explain why, based ONLY on the provided information.",
-  "",
-  "Highlight key observables in markdown bold (**…**): IP addresses, domain names, file paths,",
-  "usernames, hostnames, process names, search queries.",
-  "",
-  "Keep it short: a few paragraphs or tight bullets, not a full report. Do not invent entities or",
-  "activity the events do not contain.",
-  "",
-  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
-  JSON.stringify({ markdown: "the concise overview as raw Markdown" }, null, 2),
-].join("\n");
-
-// Per-session summary (#342): one focused call over the events of a SINGLE attacker session — a
-// contiguous run on one host — rather than the whole timeline. The events already share a host and a
-// tight time window, so the model is asked for the story of that one sitting, not a case-wide report.
-export const SESSION_SUMMARY_PROMPT = [
-  "You are a digital forensic analyst. The events below are ONE attacker session: a contiguous run",
-  "of activity on a SINGLE host, with no long gap inside it. Write a short account of what happened",
-  "during this one sitting, in Markdown.",
-  "",
-  "Cover, in this order, as flowing tight prose or bullets (no headings, no title):",
-  "- What the actor appears to have been DOING in this session, in sequence.",
-  "- The key observables — hostnames, accounts, IP addresses, file paths, process names, hashes.",
-  "- Whether the session's activity appears to have SUCCEEDED, and what in the events says so.",
-  "- What a responder should check next specifically because of this session.",
-  "",
-  "Highlight key observables in markdown bold (**…**).",
-  "",
-  "Ground EVERY statement in the supplied events. Do not invent entities, timestamps, or activity",
-  "they do not contain, and do not speculate about what happened BEFORE or AFTER this session — you",
-  "are seeing one slice of the intrusion, not the whole case. If the events are too sparse to say",
-  "what happened, say exactly that instead of guessing.",
-  "",
-  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
-  JSON.stringify({ markdown: "the session account as raw Markdown (no title heading)" }, null, 2),
-].join("\n");
-
-// Standalone narrative-timeline generator: produces a stakeholder-friendly prose story of the
-// incident. Used by `generateNarrative()` when the analyst clicks "Generate" without re-running
-// full synthesis. The same narrative is also generated as part of synthesis via SYNTHESIS_PROMPT.
-export const NARRATIVE_PROMPT = [
-  "You are a senior incident-response analyst writing a narrative timeline for ONE security incident.",
-  "Using ONLY the case evidence provided (attacker path, findings, forensic timeline), write a flowing",
-  "chronological prose story of the incident for management and non-technical stakeholders.",
-  "",
-  "Audience: decision-makers who need to understand WHAT HAPPENED and WHEN, not technical details.",
-  "Format:",
-  "- Flowing prose paragraphs — NOT bullet points.",
-  "- Chronological order, citing specific timestamps for key events.",
-  "- Third person: 'the attacker', 'the threat actor', 'the adversary'.",
-  "- Template: 'At [time], the attacker [action]. This was followed by [next step] at [time]...'",
-  "- Plain language: no ATT&CK T-codes, no hashes, no jargon. Explain tools in plain terms.",
-  "- Be honest about uncertainty: if timing or method is unclear, say 'approximately' or 'at some point'.",
-  "- 3-6 paragraphs. Each paragraph covers one phase of the intrusion.",
-  "",
-  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
-  JSON.stringify({ narrativeTimeline: "the flowing story as prose paragraphs (use \\n\\n between paragraphs)" }, null, 2),
-].join("\n");
-
-// Propose PROACTIVE Velociraptor VQL fleet-hunts from the synthesized findings (issue #57). The
-// model reads the findings / ATT&CK techniques / pivotable IOCs and emits hunts that run on EVERY
-// enrolled endpoint to find the same tradecraft elsewhere. The VQL must be deployable as-is via the
-// CLIENT-artifact hunt path (launchHunt), so the shape constraints here mirror what splitVqlStatements
-// expects: ONE self-contained VQL statement per hunt, no blank-line splits, not comment-only.
-export const HUNT_SUGGEST_PROMPT = [
-  "You are a senior DFIR threat hunter. Given ONE investigation's findings, ATT&CK techniques,",
-  "compromised assets, and indicators below, propose PROACTIVE Velociraptor VQL HUNTS that run",
-  "across the ENTIRE fleet of enrolled endpoints to find the SAME adversary tradecraft on hosts",
-  "that are not yet in scope — lateral spread, the same webshell/persistence/malware pattern,",
-  "the same C2, the same living-off-the-land technique.",
-  "",
-  "Rules:",
-  "- Hunt for the PATTERN across the fleet, do NOT merely restate a single-host finding. If a",
-  "  webshell was found in one web root, hunt every host's web roots; if a malicious service was",
-  "  installed, enumerate that service/registry value everywhere.",
-  "- When an ATTACK GRAPH is shown below, hunt for the RELATIONSHIP it reveals — the parent→child",
-  "  spawn chain, the wrote→executed file lineage, the binary/account that moved between hosts — not",
-  "  just the leaf indicator alone. A chain (e.g. winword.exe → powershell.exe → rundll32.exe) is far",
-  "  more specific than any single process name, and a binary seen on two hosts means hunt the rest.",
-  "- Pivot ONLY on the case's REAL indicators (the exact hashes, file paths, process names,",
-  "  service names, domains, IPs shown below). Do NOT invent IOCs the case does not contain.",
-  "- Each `vql` MUST be a SINGLE, self-contained, CLIENT-side Velociraptor VQL statement that runs",
-  "  on each endpoint — one `SELECT … FROM <plugin>(…) WHERE …`. Use real Velociraptor plugins,",
-  "  e.g. glob(), stat(), pslist(), Artifact.Windows.System.Services, Artifact.Windows.Sys.Users,",
-  "  read_file(), hash(), yara(), Artifact.Windows.Registry.* . Velociraptor glob() uses FORWARD",
-  "  slashes. Do NOT put a blank line inside one query and do NOT make a query only a comment.",
-  "- Use each plugin's REAL argument names. parse_evtx() takes `filename=` (a path or glob), NOT",
-  "  `files=` — e.g. parse_evtx(filename='C:/Windows/System32/winevt/Logs/Security.evtx'). For Windows",
-  "  event-log hunts PREFER Velociraptor's own artifacts (Windows.EventLogs.* / Windows.Detection.*)",
-  "  over hand-writing parse_evtx, and access fields as `System.EventID.Value` / `EventData.<Name>`.",
-  "- read_file() takes `filenames=` (a LIST), NOT `filename=` (scalar). To read one file:",
-  "  `SELECT Data FROM read_file(filenames=['C:/path/file.txt'])`. To read files found by glob,",
-  "  call it inline in the SELECT (no JOIN needed):",
-  "  `SELECT FullPath, read_file(filenames=[FullPath])[0].Data AS Content FROM glob(globs='...')`.",
-  "- VQL has NO SQL `JOIN`. Use inline function calls (e.g. read_file above) or",
-  "  `foreach(row={SELECT … FROM a()}, query={SELECT … FROM b()})` to correlate two plugins.",
-  "- hash() takes ONLY `path=` and returns an object with `.MD5` / `.SHA1` / `.SHA256` — there is NO",
-  "  `hashselect`/`algorithm`/`type` argument (inventing one fails to COMPILE, so the hunt never starts).",
-  "  Get a file's SHA256 as `hash(path=FullPath).SHA256`, computed ONCE as a column, not repeatedly in WHERE:",
-  "  `SELECT FullPath, hash(path=FullPath).SHA256 AS SHA256 FROM glob(globs='…/*.exe') WHERE SHA256 IN (…)`.",
-  "- NEVER hash with a full-disk glob like 'C:/**/*.exe' — it walks every file on the volume and times out.",
-  "  Scope globs to the directories the tradecraft uses (web roots, %TEMP%, Downloads, a service path).",
-  "- PREFER raw VQL plugins — pslist(), netstat(), glob(), stat(), read_file(), hash(), yara() — over",
-  "  `Artifact.<Name>()` references: a hallucinated Artifact.<Name> fails to COMPILE and the hunt never",
-  "  starts. Network connections → netstat(); processes → pslist(); files → glob(). Only reference an",
-  "  Artifact.<Name> you are sure exists on the server.",
-  "- For an ABSOLUTE time use timestamp(string='2025-03-14T22:00:00Z'); epoch= takes unix SECONDS (a number).",
-  "- Velociraptor VQL has NO duration-suffix literals. Do NOT write `30d`, `7h`, `2w` etc.",
-  "  Use seconds arithmetic instead: `now() - 30 * 86400` (30 days), `now() - 7 * 86400` (7 days),",
-  "  `now() - 3600` (1 hour). Wrap in `timestamp(epoch=...)` when comparing against a timestamp",
-  "  column, e.g. `WHERE Mtime > timestamp(epoch=now() - 30 * 86400)`.",
-  "- Prefer a few HIGH-SIGNAL hunts over many near-duplicates. Skip a finding if there is nothing",
-  "  fleet-wide to hunt for it.",
-  "- For each hunt set: a short `title`; a `rationale` (which finding triggered it, what the query",
-  "  looks for, and how to triage a hit); `severity` (Critical|High|Medium|Low|Info) of the",
-  "  underlying threat; `mitreTechniques` (the finding's technique ids); and `relatedFindingIds`",
-  "  (the finding ids it derives from, using the [ids] shown).",
-  "",
-  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
-  JSON.stringify({
-    suggestions: [{
-      title: "Hunt for ASPX webshells across IIS web roots",
-      rationale: "Finding f3 shows an ASPX webshell on WEB01. Sweep every endpoint's web roots for .aspx files written recently; triage any hit by author/last-write time and contents.",
-      vql: "SELECT FullPath, Mtime, Size FROM glob(globs='C:/inetpub/wwwroot/**/*.aspx') WHERE Mtime > timestamp(epoch=now() - 2592000)",
-      severity: "High",
-      mitreTechniques: ["T1505.003"],
-      relatedFindingIds: ["f3"],
-    }],
-  }, null, 2),
-].join("\n");
-
-// Issue #70 — AI-suggested Velociraptor hunts for the case's PLAYBOOK tasks. For each ENDPOINT-related
-// task the model writes one CLIENT-side VQL hunt + echoes the single host it's scoped to (chosen ONLY
-// from the provided known-endpoints list). The server decides hunt-vs-collection deterministically
-// from the observed endpoints, so the model just needs to flag endpoint-relatedness and name the host.
-export const PLAYBOOK_HUNT_PROMPT = [
-  "You are a senior DFIR threat hunter. Below is ONE investigation's Response PLAYBOOK (the analyst's",
-  "actionable checklist), the case's known ENDPOINTS, findings, ATT&CK techniques, and forensic timeline.",
-  "For each playbook task that is ABOUT ENDPOINTS (collecting from, examining, containing, or hunting on",
-  "hosts — e.g. pull event logs, enumerate persistence, find a process/file, scope lateral movement),",
-  "propose ONE Velociraptor VQL hunt that gathers the evidence the task needs.",
-  "",
-  "Rules:",
-  "- Only emit a suggestion for a task that is genuinely endpoint-related. Set `endpointRelated` true for",
-  "  those and SKIP the rest (notify legal, rotate cloud creds, draft a report, block a domain at the",
-  "  firewall, etc. are NOT endpoint tasks). It is fine to return fewer suggestions than tasks.",
-  "- `taskId` MUST be the exact id (the [bracketed] value) of the task the suggestion is for.",
-  "- `targetHost`: if the task is about exactly ONE specific endpoint, set this to that host — but ONLY a",
-  "  hostname that appears in the KNOWN ENDPOINTS list below (the server runs it as a COLLECTION on just",
-  "  that one client). If the task spans multiple hosts, or you are unsure which host, set `targetHost`",
-  "  to \"\" (the server runs it as a fleet-wide HUNT across all endpoints). Never invent a hostname.",
-  "- Each `vql` MUST be a SINGLE, self-contained, CLIENT-side Velociraptor VQL statement —",
-  "  one `SELECT … FROM <plugin>(…) WHERE …`. glob() uses FORWARD slashes. Do NOT put a blank line",
-  "  inside one query and do NOT make a query only a comment.",
-  "- STRONGLY PREFER raw VQL plugins — they ALWAYS exist: pslist(), netstat(), glob(), stat(),",
-  "  read_file(), hash(), yara(), parse_evtx(), reg_keys(). Network connections → netstat(); processes →",
-  "  pslist(); files → glob(); event logs → parse_evtx(filename='C:/Windows/System32/winevt/Logs/Security.evtx').",
-  "- You MAY reference an `Artifact.<Name>()` ONLY if <Name> appears EXACTLY in the AVAILABLE VELOCIRAPTOR",
-  "  ARTIFACTS list in the user message below. If <Name> is NOT in that list it does NOT exist on this",
-  "  server and the hunt FAILS TO COMPILE (no flow id) — use a raw plugin instead. NEVER invent an",
-  "  artifact name (e.g. Windows.EventLogs.Sysmon / .SecurityLog are NOT universal — check the list).",
-  "- Use each plugin's REAL argument names: parse_evtx() takes `filename=` (a path/glob), NOT `files=`;",
-  "  handles() takes `pid=`, NOT `process=`; read_file() takes `filenames=` (a LIST), NOT `filename=`.",
-  "  Access EVTX fields as `System.EventID.Value` / `EventData.<Name>`. To read one file:",
-  "  `SELECT Data FROM read_file(filenames=['C:/path/file.txt'])`. To read files found by glob,",
-  "  call it inline in the SELECT (no JOIN needed):",
-  "  `SELECT FullPath, read_file(filenames=[FullPath])[0].Data AS Content FROM glob(globs='...')`.",
-  "- VQL has NO SQL `JOIN`. Use inline function calls (e.g. read_file above) or",
-  "  `foreach(row={SELECT … FROM a()}, query={SELECT … FROM b()})` to correlate two plugins.",
-  "- hash() takes ONLY `path=` and returns an object with `.MD5` / `.SHA1` / `.SHA256` — there is NO",
-  "  `hashselect`/`algorithm`/`type` argument (inventing one fails to COMPILE, so the hunt never starts).",
-  "  Get a file's SHA256 as `hash(path=FullPath).SHA256`, computed ONCE as a column, not repeatedly in WHERE:",
-  "  `SELECT FullPath, hash(path=FullPath).SHA256 AS SHA256 FROM glob(globs='…/*.exe') WHERE SHA256 IN (…)`.",
-  "- NEVER hash with a full-disk glob like 'C:/**/*.exe' — it walks every file on the volume and times out.",
-  "  Scope globs to the directories the tradecraft uses (web roots, %TEMP%, Downloads, a service path).",
-  "- For an ABSOLUTE time use timestamp(string='2025-03-14T22:00:00Z'); epoch= takes unix SECONDS (a number).",
-  "- Velociraptor VQL has NO duration-suffix literals. Do NOT write `30d`, `7h`, `2w`. Use seconds",
-  "  arithmetic: `now() - 30 * 86400` (30 days), `now() - 3600` (1 hour). Wrap in `timestamp(epoch=...)`",
-  "  when comparing a timestamp column, e.g. `WHERE Mtime > timestamp(epoch=now() - 30 * 86400)`.",
-  "- Pivot on the case's REAL indicators (the exact hashes, file paths, process/service names, domains,",
-  "  IPs shown) — do NOT invent IOCs the case does not contain.",
-  "- For each: a short `title`; a `rationale` (which task it serves, what the query looks for, how to",
-  "  triage a hit); `severity` (Critical|High|Medium|Low|Info) of the underlying threat; and",
-  "  `mitreTechniques` (relevant technique ids).",
-  "",
-  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
-  JSON.stringify({
-    suggestions: [{
-      taskId: "finding:f3",
-      endpointRelated: true,
-      title: "Enumerate the malicious service on WEB01",
-      rationale: "Task asks to investigate the service-persistence finding on WEB01. Collect the host's services and flag the one whose ImagePath matches the dropped binary; triage by start type and account.",
-      vql: "SELECT Name, DisplayName, PathName, StartMode FROM Artifact.Windows.System.Services() WHERE PathName =~ 'evil\\\\.exe'",
-      targetHost: "WEB01",
-      severity: "High",
-      mitreTechniques: ["T1543.003"],
-    }],
-  }, null, 2),
-].join("\n");
-
-// Hypothesise attacker actions for TIMELINE GAPS (issue #96). The deterministic gap detector has
-// already flagged suspiciously silent periods; the model reads each gap's bounding context — the
-// events just BEFORE the silence and just AFTER — and infers what the attacker most likely did during
-// the hole (e.g. cleared the log to hide credential dumping, disabled EDR before lateral movement).
-// It is grounded ONLY in the surrounding events; it does NOT invent activity. It also names which
-// SHADOW ARTIFACTS (from the catalog ids in the user message) would best reconstruct each window —
-// the deterministic catalog supplies the actual collection VQL, so the model only ranks relevance.
-export const GAP_HYPOTHESIS_PROMPT = [
-  "You are a senior DFIR analyst reasoning about COVERAGE GAPS in ONE investigation's forensic timeline.",
-  "Each gap below is a stretch where logging went silent — a COMPLETE gap (every source dark) is the",
-  "classic signature of cleared Windows Event Logs, a stopped collector/auditd, or disabled EDR. For",
-  "EACH gap, hypothesise what the attacker most likely did DURING the silence, reasoning from the events",
-  "immediately BEFORE the gap (what they were doing) and immediately AFTER (the state when logging",
-  "resumed).",
-  "",
-  "Rules:",
-  "- Ground every hypothesis ONLY in the surrounding events shown. Do NOT invent specific hosts, files,",
-  "  or accounts the context does not mention. If the surrounding events are too sparse to say anything,",
-  "  give a low confidence and say the gap is unexplained.",
-  "- Prefer the explanation that fits the tradecraft: a complete silence right after initial access often",
-  "  hides discovery/credential-access/defense-evasion (clearing logs to cover the next step); a gap",
-  "  bracketed by a logon and later persistence often hides lateral movement or staging.",
-  "- `gapId` MUST be the exact [gap-N] id shown for the gap the hypothesis is about. Emit at most one",
-  "  hypothesis per gap. It is fine to skip a gap that is plainly benign (e.g. an expected overnight quiet).",
-  "- `hypothesis`: 2-4 sentences naming the most probable attacker activity and WHY it fits the context.",
-  "- `attackerActions`: a few concrete candidate actions that would produce this exact gap.",
-  "- `confidence`: 0-100, honest — sparse context or an equally-likely benign explanation means LOW.",
-  "- `severity`: Critical|High|Medium|Low|Info — how serious the hypothesised activity would be.",
-  "- `mitreTechniques`: ATT&CK ids for the hypothesised actions (e.g. T1070.001 for cleared event logs).",
-  "- `recommendedArtifactIds`: from the SHADOW ARTIFACTS list in the user message, the ids whose data",
-  "  would best confirm THIS hypothesis (e.g. prefetch/amcache/shimcache/bam for execution, usn-journal/",
-  "  mft/lnk-files for file activity, srum for exfiltration). Use ONLY ids from that list.",
-  "",
-  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
-  JSON.stringify({
-    hypotheses: [{
-      gapId: "gap-1",
-      hypothesis: "The Security log was cleared immediately after the initial RDP logon and before the service-install seen on resume, consistent with the attacker wiping logs to hide credential access and staging during the silence.",
-      attackerActions: ["Cleared the Windows Security event log (wevtutil cl / EventLog API)", "Dumped LSASS or ran a discovery tool while logging was off"],
-      confidence: 55,
-      severity: "High",
-      mitreTechniques: ["T1070.001", "T1003.001"],
-      recommendedArtifactIds: ["prefetch", "amcache", "usn-journal", "srum"],
-    }],
-  }, null, 2),
-].join("\n");
-
-// Memory-forensics "Next-Step" agent (issue #101). The case already has Volatility 3 / Rekall output
-// imported as forensic events (the process tree, network connections, malfind injected code, command
-// lines, services, modules). Read that memory evidence, identify the ANOMALIES, and propose the EXACT
-// next Volatility 3 command the analyst should run to dig deeper. The agent CONSUMES the enumeration
-// (it does not re-implement Volatility) — it reasons over the rows and recommends the next plugin.
-export const MEMORY_NEXTSTEP_PROMPT = [
-  "You are a senior memory-forensics analyst guiding an ITERATIVE Volatility 3 investigation. Below is",
-  "the memory evidence ALREADY imported from a RAM image (Volatility 3 / Rekall output): the process",
-  "tree (process name, PID, PPID, parent name, start time, command line), network connections, malfind",
-  "(executable/injected private memory), command lines, services, and loaded modules. Identify the",
-  "ANOMALIES and, for each, propose the EXACT next Volatility 3 command the analyst should run to dig in.",
-  "",
-  "What counts as an anomaly (reason from the evidence shown, do NOT invent processes/PIDs):",
-  "- Process-tree masquerading / wrong parentage: svchost.exe NOT parented by services.exe; lsass.exe,",
-  "  csrss.exe, services.exe, wininit.exe with the wrong/absent parent; an unparented process; a system",
-  "  binary running from a non-system path; a user app spawning cmd.exe/powershell.exe.",
-  "- Injected/executable private memory (malfind hits) → confirm whether it is real injection or benign.",
-  "- Suspicious or external network connections owned by an unexpected process (possible C2/beacon).",
-  "- LOLBin / encoded-PowerShell / unusual command lines.",
-  "- A persistence-looking service or an unsigned/odd module.",
-  "",
-  "APPLY FALSE-POSITIVE AWARENESS (this is the most important part — every malfind hit is ingested as",
-  "High, but many RWX/executable-private-memory regions are BENIGN). Before proposing a dig-in step, ask",
-  "whether the hit is expected for that process:",
-  "- Security/AV engines legitimately use RWX: MsMpEng.exe (Microsoft Defender), MpDefenderCoreService,",
-  "  and third-party AV/EDR. These are the #1 malfind false positive.",
-  "- .NET/CLR and other JIT compilers emit RWX: powershell.exe, processes hosting the CLR, and",
-  "  JavaScript/Java/Lua JITs (browsers — chrome/msedge/firefox, node, java). RWX here is normal JIT.",
-  "- Some legitimate packers/installers and SearchHost.exe/Search/UI shell processes also show RWX.",
-  "When the malfind hit is on such a process AND nothing else about it is anomalous (correct image path,",
-  "correct parent, no suspicious cmdline/connection), SAY SO in the `anomaly`/`rationale`, set `severity`",
-  "to Low or Info, and make the next step a quick LEGITIMACY CONFIRMATION — `windows.cmdline` /",
-  "`windows.dlllist` to verify the image path, signer, and loaded modules — rather than dumping every",
-  "region. Reserve High/Critical and a real dig-in for genuinely unexpected processes, wrong parentage,",
-  "bad paths, or malfind correlated with a suspicious connection/command line.",
-  "",
-  "Rules:",
-  "- Each `command` MUST be a single, real, copy-pasteable Volatility 3 command. Use `vol -f <image>`",
-  "  as the prefix (the analyst substitutes their image path for <image>) followed by a REAL Volatility 3",
-  "  plugin and its REAL options, e.g.:",
-  "    vol -f <image> windows.malfind --pid 1234",
-  "    vol -f <image> windows.dlllist --pid 1234",
-  "    vol -f <image> windows.cmdline --pid 1234",
-  "    vol -f <image> windows.handles --pid 1234",
-  "    vol -f <image> windows.netscan",
-  "    vol -f <image> windows.pstree",
-  "    vol -f <image> windows.getsids --pid 1234",
-  "    vol -f <image> windows.svcscan",
-  "  Use Linux/Mac plugin names (linux.* / mac.*) instead if the evidence is clearly from that OS.",
-  "  Use the REAL plugin/option names — do NOT invent plugins or flags, and do NOT use Volatility 2",
-  "  syntax (no `--profile`, no `vol.py -f mem.raw pslist`-style v2 plugin names).",
-  "- STRONGLY PREFER commands that produce a TABLE the analyst can paste/import straight back into this",
-  "  tool (malfind, cmdline, handles, dlllist, netscan, pstree, svcscan, getsids, privileges, …). Do NOT",
-  "  add `--dump` and do NOT suggest a plain `windows.dumpfiles`/`windows.procdump` as the step UNLESS",
-  "  dumping is genuinely the right next move — a dump writes a RAW BINARY .dmp/.exe to disk, which is",
-  "  NOT something this tool can ingest. When you DO recommend a dump, the `rationale` MUST say the .dmp",
-  "  is for OFFLINE analysis (YARA/`capa`/`strings`/upload to a malware sandbox) and that the analyst",
-  "  imports THOSE results back (this tool ingests sandbox reports) — the .dmp itself is not re-imported.",
-  "- PREFER suggesting plugins that have NOT been run yet (the user message lists the already-imported",
-  "  plugins) when they would advance the investigation — the point is the NEXT step, not re-running",
-  "  what is already on the timeline. Pivot on a SPECIFIC PID/process from the evidence wherever the",
-  "  plugin takes `--pid`; set `pid` to that PID.",
-  "- Prefer a few HIGH-SIGNAL next steps over many near-duplicates. If nothing in the evidence looks",
-  "  anomalous, it is fine to return fewer (or no) suggestions.",
-  "- For each: a short `anomaly` (the observation that triggered it, naming the real process/PID); the",
-  "  `command`; the `plugin` it runs (e.g. windows.malfind); a `rationale` (why run it + how to triage",
-  "  what it returns); `severity` (Critical|High|Medium|Low|Info) of the underlying anomaly;",
-  "  `pid` (the targeted PID, or \"\"); and `mitreTechniques` (relevant ATT&CK ids, e.g. T1055 for injection).",
-  "",
-  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
-  JSON.stringify({
-    suggestions: [{
-      anomaly: "svchost.exe (PID 1234) is parented by explorer.exe (PID 4500), not services.exe — classic masquerading.",
-      command: "vol -f <image> windows.malfind --pid 1234",
-      plugin: "windows.malfind",
-      rationale: "A mis-parented svchost is a strong injection/masquerade signal. malfind dumps executable private memory in the process; triage any MZ/shellcode region by yara-scanning the dump and pivot on its imports.",
-      severity: "High",
-      pid: "1234",
-      mitreTechniques: ["T1055", "T1036.005"],
-    }],
-  }, null, 2),
-].join("\n");
-
-// Translate a free-text analyst request into a runnable hunting query per platform (issue #100).
-// The analyst describes the activity in plain English; the model maps that intent onto each requested
-// platform's REAL schema (the per-platform reference is supplied in the user message). EPHEMERAL like
-// ask()/suggestHunts(): no state change. The Velociraptor query is deployable as-is via launchHunt, so
-// the VQL shape constraints mirror what splitVqlStatements expects (one self-contained statement).
-export const QUERY_TRANSLATE_PROMPT = [
-  "You are a senior DFIR detection engineer. An analyst will give you a request in PLAIN ENGLISH describing",
-  "the activity they want to find in their logs (e.g. \"PowerShell downloading a file and then executing it\",",
-  "\"outbound RDP from this host\", \"new local administrator accounts\"). Translate that intent into a runnable",
-  "hunting query for EACH target platform listed in the user message, grounded in that platform's REAL schema",
-  "(the tables / plugins / field names given for it).",
-  "",
-  "Rules:",
-  "- Emit one entry per TARGET PLATFORM key shown in the user message — use the exact key (velociraptor,",
-  "  defender, elastic, splunk, sigma, yara, suricata). Do NOT invent platforms or emit one not requested.",
-  "- Ground every query in the platform's REAL schema shown for it — its actual tables/plugins/field names.",
-  "  Do NOT invent table or field names; prefer the canonical fields listed.",
-  "- Capture the analyst's FULL intent, including sequencing/relationships where the platform allows it (e.g.",
-  "  a parent→child process relationship, \"download THEN execute\"). When a platform can't express a relation,",
-  "  approximate with the closest field filters and note the limitation in `caveats`.",
-  "- If the request references a specific case entity (\"this host\", \"that IP/hash\"), use the matching value",
-  "  from the PIVOTABLE INDICATORS list — do NOT invent indicators the case does not contain.",
-  "- Velociraptor `query` MUST be a SINGLE, self-contained, CLIENT-side VQL statement — one",
-  "  `SELECT … FROM <plugin>(…) WHERE …`. glob() uses FORWARD slashes; VQL has NO SQL JOIN (use foreach() or",
-  "  inline calls) and NO duration literals (use now() - N * 86400). Do NOT put a blank line inside one query.",
-  "- KQL / ES|QL / SPL queries must be directly runnable (piped where idiomatic). Sigma is a YAML detection",
-  "  rule; Suricata is `alert … (msg:…; …; sid:9000001; rev:1;)`; YARA targets FILE CONTENT.",
-  "- If a platform genuinely cannot express the request (e.g. YARA or Suricata for a pure process-behavior",
-  "  request with no file/network indicator), set `notApplicable` true, leave `query` empty, and explain why",
-  "  in `caveats` — do NOT force a meaningless query.",
-  "- For each entry: a short `label`; the `query`; an `explanation` (how it captures the request + what a hit",
-  "  looks like); optional `caveats` (assumptions, field-mapping notes, what to verify before trusting hits).",
-  "- Also return a one-sentence `interpretation` of how you understood the request, so the analyst can confirm",
-  "  intent before running anything.",
-  "",
-  "Return ONLY raw JSON (no markdown fences) with EXACTLY this shape:",
-  JSON.stringify({
-    interpretation: "Find PowerShell processes that download a file and then execute it.",
-    queries: [{
-      platform: "velociraptor",
-      label: "PowerShell download-and-execute (live processes)",
-      query: "SELECT Pid, Ppid, Name, CommandLine, Exe FROM pslist() WHERE Name =~ '(?i)powershell' AND CommandLine =~ '(?i)(DownloadString|DownloadFile|Invoke-WebRequest|iwr|curl|wget)' AND CommandLine =~ '(?i)(Invoke-Expression|iex|Start-Process|-enc)'",
-      explanation: "Lists running PowerShell processes whose command line shows BOTH a download primitive and an execution primitive; a hit is one process doing both.",
-      caveats: "Live process list only — pair with EID 4104 script-block logs / Sysmon 1 for historical coverage.",
-      notApplicable: false,
-    }, {
-      platform: "defender",
-      label: "PowerShell download then execute",
-      query: "DeviceProcessEvents\n| where FileName =~ \"powershell.exe\"\n| where ProcessCommandLine has_any (\"DownloadString\",\"DownloadFile\",\"Invoke-WebRequest\",\"iwr\",\"curl\",\"wget\")\n| where ProcessCommandLine has_any (\"Invoke-Expression\",\"iex\",\"Start-Process\",\"-enc\")",
-      explanation: "Returns PowerShell process events whose command line contains both a download and an execution primitive.",
-      caveats: "Tune the keyword lists; join DeviceNetworkEvents to confirm the download egress.",
-      notApplicable: false,
-    }],
-  }, null, 2),
-].join("\n");
-
-export const getSystemPrompt = (): string => resolvePrompt("SYSTEM", SYSTEM_PROMPT);
-export const getCsvPrompt = (): string => resolvePrompt("CSV", CSV_SYSTEM_PROMPT);
-export const getLogPrompt = (): string => resolvePrompt("LOG", LOG_SYSTEM_PROMPT);
-export const getSynthesisPrompt = (): string => resolvePrompt("SYNTH", SYNTHESIS_PROMPT);
-export const getObservePrompt = (): string => resolvePrompt("OBSERVE", OBSERVE_PROMPT);
 
 /** What one deep-pass run did, for the analyst and the route response. */
 export interface DeepPassResult {
@@ -1457,59 +257,6 @@ export interface DeepPassResult {
   batchesFailed: number;  // batches whose response never parsed — that slice went unread
   observations: number;   // observations that survived sanitising
 }
-
-export const getAskPrompt = (): string => resolvePrompt("ASK", ASK_PROMPT);
-export const getExecSummaryPrompt = (): string => resolvePrompt("EXEC", EXEC_SUMMARY_PROMPT);
-export const getNarrativePrompt = (): string => resolvePrompt("NARRATIVE", NARRATIVE_PROMPT);
-export const getHuntSuggestPrompt = (): string => resolvePrompt("HUNTS", HUNT_SUGGEST_PROMPT);
-export const getPlaybookHuntPrompt = (): string => resolvePrompt("PBHUNTS", PLAYBOOK_HUNT_PROMPT);
-export const getGapHypothesisPrompt = (): string => resolvePrompt("GAPHYP", GAP_HYPOTHESIS_PROMPT);
-export const getMemoryNextStepPrompt = (): string => resolvePrompt("MEMNEXT", MEMORY_NEXTSTEP_PROMPT);
-export const getQueryTranslatePrompt = (): string => resolvePrompt("QUERYXLATE", QUERY_TRANSLATE_PROMPT);
-export const getReconcilePrompt = (): string => resolvePrompt("RECONCILE", RECONCILE_PROMPT);
-export const getExplainEventPrompt = (): string => resolvePrompt("EXPLAIN", EXPLAIN_EVENT_PROMPT);
-export const getRemediationPrompt = (): string => resolvePrompt("REMEDIATION", REMEDIATION_PROMPT);
-export const getFpSimilarityPrompt = (): string => resolvePrompt("FPSIMILARITY", FP_SIMILARITY_PROMPT);
-export const getHypothesisReviewPrompt = (): string => resolvePrompt("HYPREVIEW", HYPOTHESIS_REVIEW_PROMPT);
-export const getStarredReportPrompt = (): string => resolvePrompt("STARREDREPORT", STARRED_REPORT_PROMPT);
-export const getViewSummaryPrompt = (): string => resolvePrompt("VIEWSUMMARY", VIEW_SUMMARY_PROMPT);
-export const getSessionSummaryPrompt = (): string => resolvePrompt("SESSIONSUMMARY", SESSION_SUMMARY_PROMPT);
-
-export const IMPORTER_PROMPT = [
-  "You are writing a DECLARATIVE IMPORTER DEFINITION for the DFIR Companion. Output ONLY a single",
-  "JSON object conforming to the schema below — no prose, no markdown fences.",
-  "",
-  "The importer tells the Companion how to (1) RECOGNIZE a file by its shape, and (2) MAP each row",
-  "into a forensic event + IOCs. It is pure data; no code runs.",
-  "",
-  "SCHEMA (all fields):",
-  "- id: kebab-case unique id (not a built-in name). label: human name. version: 1.",
-  "- match: how to detect the file. format: csv|json|ndjson|auto. For CSV use requireHeaders (all",
-  "  present) / anyHeaders (>=1 present). For JSON use requireKeys / anyKeys / keyEquals {key: regex}.",
-  "  Optional filenamePattern (regex). priority: lower = tried earlier (default 100).",
-  "  Every regex is ReDoS-checked and REJECTED if it can backtrack catastrophically: no quantifier",
-  "  inside a repeated group ((a+)+, (\\d{1,3})+), no repeated group whose alternatives can start",
-  "  with the same character ((a|ab)+), no adjacent open-ended repeats (.*.*). Keep them simple.",
-  "- map: timestamp {from:[cols], format:auto|iso|epoch_s|epoch_ms} (REQUIRED); description: a",
-  "  template string with {{ColumnName}} placeholders (REQUIRED); severity: a fixed level OR",
-  "  {from:[col], map:{value:Level}, default:Level}; asset (host), user (account), processName,",
-  "  parentName, sha256, md5, path, srcIp, dstIp, port: each {from:[cols], transform?}; mitre:",
-  "  {from:[col]} (parses Txxxx) or {fixed:[...]}; iocs: list of {type,from:[cols]} or",
-  "  {autoExtract:[cols]}. Levels: Critical|High|Medium|Low|Info. transforms: trim|lowercase|",
-  "  basename|cleanIp|defang|refang.",
-  "- options: { aggregate:true, minSeverity?, maxEvents?, maxIocs? }.",
-  "",
-  "WORKED EXAMPLE (a valid importer):",
-  JSON.stringify(EXAMPLE_IMPORTER_SPEC, null, 2),
-  "",
-  "Now write an importer for THIS file. Paste a representative sample of your exported file below,",
-  "then I will return one JSON importer definition that maps it for the DFIR Companion.",
-  "",
-  "FILE SAMPLE:",
-  "<<< paste a few representative rows/records of your export here >>>",
-].join("\n");
-
-export const getImporterPrompt = (): string => resolvePrompt("IMPORTGEN", IMPORTER_PROMPT);
 
 // Result of the two view-scoped AI summaries (starred report / view summary). `eventCount` is the
 // full deduplicated match; `usedEvents` what actually fit the AI input budget.
